% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{alignData}
\alias{alignData}
\title{align genome data at specified coordinates (e.g. TSS)}
\usage{
alignData(coors, data, dst = 500, chrS)
}
\arguments{
\item{coors}{genome positions (position, chromosome, strand)
at which data will be aligned}

\item{data}{genome data to be aligned; NOTE, that currently this
is required to be fully expanded matrix covering each position
(TODO: allow non-expanded data)}

\item{dst}{upstream/downstream length to be aligned
(TODO: allow different upstream and downstream ranges)}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information, see function \code{\link{getChrSum}}}
}
\description{
align genome data at specified coordinates (e.g. TSS)
}
