% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{annotateTarget}
\alias{annotateTarget}
\title{wrapper around \test{\code{segmentOverlap}}, used to 
annotate the target set by a column in the query set}
\usage{
annotateTarget(query, target, col, prefix, details = FALSE,
  duplicates = "collapse")
}
\arguments{
\item{query}{the query set of segments (genomic intervals)}

\item{target}{the target set of segments (genomic intervals)}

\item{col}{column names to copy from target to it's best match in query}

\item{prefix}{column name prefix for the copied columns}

\item{details}{set to \code{TRUE} to add details of the used
match, i.e., union and intersect, and relative position of the query
to the matching targets}
}
\description{
wrapper around \test{\code{segmentOverlap}}, used to 
annotate the target set by a column in the query set
}

