% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{annotationOverlap}
\alias{annotationOverlap}
\title{analyze internal overlaps in a TRUE/FALSE feature annotation table,
where rows are features (eg. genes) and columns are annotation
terms. Columns must be named with the annotation terms/IDs.}
\usage{
annotationOverlap(x)
}
\arguments{
\item{x}{a logical (TRUE/FALSE) feature annotation table}
}
\description{
analyze internal overlaps in a TRUE/FALSE feature annotation table,
where rows are features (eg. genes) and columns are annotation
terms. Columns must be named with the annotation terms/IDs.
}
