% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{arno}
\alias{arno}
\title{Color Ramp Palette `arno`}
\usage{
arno(n)
}
\arguments{
\item{n}{number of colors to return}
}
\description{
generates colors similar to viridis::inferno but with a more visible
yellow, taken from viridis::viridis. Named after Arno because it
was his birthday.
}
