% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{ash}
\alias{ash}
\title{asinh data transformation}
\usage{
ash(x)

ash(x)
}
\arguments{
\item{x}{data to be transformed}

\item{x}{data to be transformed}
}
\value{
log(x+sqrt(x^2+1))

log(x+sqrt(x^2+1))
}
\description{
asinh trafo, an alternative to log transformation that has less
(compressing) effects on the extreme values (low and high values),
and naturally handles negative numbers and 0

asinh trafo, an alternative to log transformation that has less
(compressing) effects on the extreme values (low and high values),
and naturally handles negative numbers and 0
}
