% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{ashaxis}
\alias{ashaxis}
\title{Draw axis for arcsinh transformed data}
\usage{
ashaxis(side, at = c(-10^(7:0), 0, 10^(0:7)), lat = 0:100, ...)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is
    to be drawn on.  The axis is placed as follows: 1=below,
    2=left, 3=above and 4=right.}

\item{at}{the points at which tick-marks are to be drawn.  Non-finite
    (infinite, \code{NaN} or \code{NA}) values are omitted.  By default
    (when \code{NULL}) tickmark locations are computed, see
    \sQuote{Details} below.}

\item{lat}{as argument \code{at} of \link[graphics:axis]{axis} but
as the exponents of \code{log(y, base=base)}}

\item{...}{parameters passed on to \link[graphics:axis]{axis}}
}
\description{
Draw axis for arcsinh transformed data
}
