% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{bed2coor}
\alias{bed2coor}
\title{parse a .bed format file}
\usage{
bed2coor(file, header = c("chr", "start", "end", "name", "score"))
}
\arguments{
\item{file}{a file in bed format}

\item{header}{header names for the parsed bed file}
}
\description{
UNTESTED: Parses a bed file with 5 (or more) columns into the
segmenTools genomic interval format.
}
