% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{ci95}
\alias{ci95}
\title{calculate 95% confidence intervals for the given
data vector using a t-distribution}
\usage{
ci95(x, na.rm = FALSE)
}
\arguments{
\item{x}{data vector for which the 95% confidence interval
will be calculated}

\item{na.rm}{remove NA values before calculation}
}
\description{
calculate 95% confidence intervals for the given
data vector using a t-distribution
}

