% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{classify_trend}
\alias{classify_trend}
\title{Classify trends in short vectors.}
\usage{
classify_trend(y, x, min_n = 4, smooth_df = 4, slope_tol = 0.01)
}
\arguments{
\item{y}{y-values.}

\item{x}{x-values.}

\item{min_n}{minimum number of data points to be considered.}

\item{smooth_df}{degrees of freedom for
\link[stats:smooth.spline]{smooth.spline}.}

\item{slope_tol}{threshold for ignoring small slopes.}
}
\description{
Classifies each vector into increasing, decreasing, flat, unimodal
(one peak or trough), or complex; and nodata (n<4), or nofit (fit
error), using \link[stats:smooth.spline]{smooth.spline} error
classes. Note that this was generated with the help of chatGPT,
and has not been optimized.
}
