% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{clusterCluster}
\alias{clusterCluster}
\title{calculates overlaps between two clusterings}
\usage{
clusterCluster(
  cl1,
  cl2,
  na.string = "na",
  cl1.srt,
  cl2.srt,
  alternative = c("greater")
)
}
\arguments{
\item{cl1}{clustering 1; a vector of cluster associations or an
object of class "clustering" by segmenTier's
\code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}}

\item{cl2}{clustering 2; see argument \code{cl1}}

\item{na.string}{replace NA or empty strings by `<na.string>'}

\item{cl1.srt}{optional cluster sorting of clustering 1}

\item{cl2.srt}{optional cluster sorting of clustering 2}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of `"greater"' to calculate enrichment (default),
`"less"' to calculate deprivation, or `"two.sided"' to report the more
more signficant p-value of both}
}
\description{
Calculates mutual overlaps between two clusterings of the same data set
using hypergeometric distribution statistics for significantly
enriched or deprived mutual overlaps. The resulting tables can
be plotted with \code{\link{plotOverlaps}} and sorted along one
axis by signficance with \code{\link{plotOverlapss}}.
TODO: specify wich cl will be rows/columns and to which
percent refers to
}
