% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{clusterColors}
\alias{clusterColors}
\title{get color vector for clustered features}
\usage{
clusterColors(cset, expand = TRUE, ...)
}
\arguments{
\item{cset}{a structure of class 'clustering' as returned by
segmenTier's \code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}}

\item{expand}{logical indicating whether colors should be expanded
to all data}

\item{...}{arguments to cluster selection with function
\code{\link{selected}}, e.g., cluster number K}
}
\description{
Retrieves a color vector for clustered features from
segmenTier's clustering object. Default is to return colors for the
pre-selected clustering via option \code{K} to function
\code{\link{selected}}. Not to mix up with segmenTier's
function \code{\link[segmenTier:colorClusters]{colorClusters}} which
assigns a cluster coloring scheme.
}
