% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{clusterFlow}
\alias{clusterFlow}
\title{Alluvial plot for cluster matrices.}
\usage{
clusterFlow(
  clusters,
  srt,
  cls.col,
  col = "gray",
  border = 0,
  layer,
  hide = FALSE,
  alpha = 0.5,
  gap.width = 0.05,
  xw = 0.1,
  cw = 0.1,
  blocks = TRUE,
  axis_labels = NULL,
  cex = par("cex"),
  font = par("font"),
  cex.axis = par("cex.axis")
)
}
\arguments{
\item{clusters}{a matrix of different clusterings (columns).}

\item{srt}{a list of cluster orders for each column in \cite{clusters}.}

\item{cls.col}{a list of cluster colors used for boxes at start and end
of flow; must have the same order as clusters in \code{srt}.}
}
\description{
Adapted from the code of \cite{alluvial::alluvial}.
}
