% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{clusterTimeseries2}
\alias{clusterTimeseries2}
\title{Cluster a processed time-series with k-means or flowClust}
\usage{
clusterTimeseries2(tset, K = 16, method = "flowClust", selected,
  parameters = c(iter.max = 1e+05, nstart = 100, B = 500, tol = 1e-05, lambda
  = 1, nu = 4, nu.est = 0, trans = 1, merge = FALSE, randomStart = 0),
  nui.thresh = -Inf, verb = 1, ...)
}
\arguments{
\item{tset}{a timeseries processed by \code{\link[segmenTier:processTimeseries]{processTimeseries}}}

\item{K}{selected cluster numbers, the argument \code{centers}
of \code{\link[stats:kmeans]{kmeans}}}

\item{method}{string specifying the clustering algorithm to use}

\item{selected}{a pre-selected cluster number  which is then
used as a start clustering for \code{flowMerge} (if option
\code{merge==TRUE})}

\item{parameters}{named vector of parameters for clustering algorithms,
currently ALL required parameters MUST be specified}

\item{nui.thresh}{threshold correlation of a data point to a cluster
center; if below the data point will be added to nuissance cluster 0}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}

\item{...}{further parameters to \code{flowClust} or \code{\link[stats:kmeans]{kmeans}}}
}
\description{
A wrapper for clustering a time-series object \code{tset} provided by
\code{\link[segmenTier:processTimeseries]{processTimeseries}},
where specifically the DFT of a time-series and requested data
transformation were calculated. The clustering is performed
on the \code{tset$dat} matrix by \code{\link[stats:kmeans]{kmeans}} or
model-based by packages \pkg{flowClust} & \pkg{flowMerge}.
}
\details{
This function attempts to combine the previous separate clustering
wrappers from package \code{segmenTier}, \code{\link[segmenTier:flowclusterTimeseries]{flowclusterTimeseries}} and k-mean's based \code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}, the latter of which is used for
the segmenTier algorithm. Please see the corresponding help files for
details on the clustering parameters in argument \code{parameters}.
}
