% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{col2alpha}
\alias{col2alpha}
\title{R colors with alpha value.}
\usage{
col2alpha(col = "red", alpha = 1)
}
\arguments{
\item{col}{and R color specifications such as numeric string colors.}

\item{alpha}{the desired alpha (opaqueness) level from 0
(transparent) to 1 (opaque).}
}
\description{
R colors with alpha value.
}
