% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{collapsePositions}
\alias{collapsePositions}
\title{collapse adjacent genome positions into segments}
\usage{
collapsePositions(x, dist = 1)
}
\arguments{
\item{x}{ordered chromosome coordinates, a matrix or data.frame
with columns "chr" and "coor"}

\item{dist}{maximal distance between loci to be collapsed}
}
\description{
collapse adjacent genome positions into segments
}
