% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osciTools.R
\name{complex2degree}
\alias{complex2degree}
\title{convert complex numbers to degrees}
\usage{
complex2degree(x)
}
\arguments{
\item{x}{a vector or matrix of complex numbers}
}
\description{
convert complex number representation of polar coordinates
(such as provided by \code{\link[stats]{mvfft}} or segmenTools'
wrapper \code{\link{get_fft}}
to degrees, and map to 0-360 representing the first cycle
in an experiment
}
\seealso{
\code{\link{get_fft}}
}
