% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{coor2bed}
\alias{coor2bed}
\title{export internal coordinate format to bed file format}
\usage{
coor2bed(
  coor,
  file,
  coors = c(chr = "chr", start = "start", end = "end", strand = "strand"),
  reverse = c("-", -1),
  name,
  score,
  prefix,
  verb = 1
)
}
\arguments{
\item{coor}{genome coordinates}

\item{file}{optional name for bed output file}

\item{coors}{column names of coordinate columns}

\item{reverse}{reverse strand characters}

\item{name}{column name for bed file name column (column 4)}

\item{score}{column name for bed file score column (column 5)}

\item{prefix}{prefix to be added to the name and score columns (column 4,5),
used by segmenTools interface to bedtools for unique names.}

\item{verb}{verbosity level, 0: silent}
}
\description{
export internal coordinate format to bed file format
}
\seealso{
bed2coor
}
