% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{coor2index}
\alias{coor2index}
\title{convert chromosome coordinates to continuous index}
\usage{
coor2index(
  features,
  chrS,
  chrMap,
  cols = c("start", "end", "coor"),
  chrCol = "chr",
  strandCol = "strand",
  reverse = c("-", -1),
  circular = FALSE
)
}
\arguments{
\item{features}{a table of chromosome features that must contain
the chromosome number (option \code{chrCol}), one or more chromosome
positions (option \code{cols}) and strand information (column
 \code{strandCol}).}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information; simply the cumulative lengths of ordered chrosomes,
see function \code{\link{getChrSum}}}

\item{chrMap}{a vector of chromosome names using \code{features}' chromosome
column, in the same order as \code{chrS}}

\item{cols}{name of the columns giving coordinates that will be mapped
to continuous index}

\item{chrCol}{name of the column that gives the chromosome number}

\item{strandCol}{name of the column that gives forward/reverse strand
information}

\item{reverse}{a vector of possible reverse strand indicators}

\item{circular}{suppresses re-sorting to start < end for circular chromosomes}
}
\description{
convert chromosome coordinates to continuous index
}
