% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{coorMerge}
\alias{coorMerge}
\title{merge genomic coordinates by type}
\usage{
coorMerge(coor, type, verb = 1)
}
\arguments{
\item{coor}{genomic coordinates.}

\item{type}{column name of the types which should be merged
separately.}

\item{verb}{verbosity level, 0 for silent}
}
\description{
merge genomic coordinates by type
}
