% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{coorSort}
\alias{coorSort}
\title{sort genomic coordinates}
\usage{
coorSort(coor, by.strand = FALSE, reverse = c("-1", "-"))
}
\arguments{
\item{coor}{genomic coordinates.}

\item{by.strand}{split by strand.}

\item{reverse}{reverse strand identifiers, all others are
considered to stem from the forward strand.}
}
\description{
Sorts genomic coordinates (genomic intervals) along chromosomes
}
