% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{coorUpstream}
\alias{coorUpstream}
\title{get upstream or downstream range}
\usage{
coorUpstream(coor, upstream, reverse = c("-1", "-"))
}
\arguments{
\item{coor}{genomic coordinates.}

\item{upstream}{usptream (<0) or downstream (>0) range of the intervals
in \code{coor}.}

\item{reverse}{reverse strand identifiers, all others are
considered to stem from the forward strand.}
}
\description{
Returns upstream (upstream<0) or downstream (upstream>0) ranges of
the passed features. The reported ranges start at the start or end
coordinates, in a strand-specific way.
}
