% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Plot correlation matrices.}
\usage{
cor_matrix(
  x,
  n = 100,
  cols = colorRampPalette(c("blue", "white", "red"))(n),
  breaks = seq(-1, 1, length.out = n + 1),
  min.cor = 0.5,
  round.cor = 1,
  diagonal = FALSE,
  upper = FALSE,
  lower = TRUE,
  ...
)
}
\arguments{
\item{x}{a correlation matrix, values in [0,1].}

\item{n}{number of colors.}

\item{cols}{color vector (length n).}

\item{breaks}{breaks (length n+1).}

\item{min.cor}{minimal absolute correlation above which text
(correlation) is plotted in white text.}

\item{round.cor}{number of decimals to which the correlation (text)
should be rounded.}

\item{upper}{show the upper triangle of the correlation matrix.}

\item{lower}{show the lower triangle of the correlation matrix.}

\item{...}{arguments passed on to \link{image_matrix}.}

\item{diagnol}{show the diagonal (should be correlations of 1)
instead of blanking it.}
}
\description{
Plot correlation matrices.
}
