% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{dateaxis}
\alias{dateaxis}
\title{Nicely formated date axis over three date ranges.}
\usage{
dateaxis(side = 1, xlim)
}
\arguments{
\item{side}{axis side.}

\item{xlim}{x-axis limits (as.POSIXct), used to calculate  appropriate ticks.}
}
\description{
Nicely formated date axis over three date ranges.
}
