% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osciTools.R
\name{degree2complex}
\alias{degree2complex}
\title{convert phase (degree) to polar coordinates}
\usage{
degree2complex(phase, amplitude = 1, unit = "degree")
}
\arguments{
\item{phase}{phase in degree}

\item{amplitude}{amplitude}

\item{unit}{"degree" or "rad"}
}
\description{
convert phase (degree) to polar coordinates
}
\seealso{
\code{\link{complex2degree}}
}
