% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{dense2d}
\alias{dense2d}
\title{2D density heatmap plot}
\usage{
dense2d(x, y, pch = 20, ...)
}
\arguments{
\item{x}{x-coordinates}

\item{y}{y-coordinates}

\item{pch}{\code{pch} argument to plot}

\item{...}{arguments to plot (hint:cex can be useful)}
}
\description{
copied from Josh O'Brien posted at
https://stackoverflow.com/questions/17093935/r-scatter-plot-symbol-color-represents-number-of-overlapping-points/17096661#17096661
}
