% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{dotplot}
\alias{dotplot}
\title{dotplot}
\usage{
dotplot(
  x,
  p.min = 0.01,
  dot.sze = c(0.3, 2),
  value = c("overlap", "count", "statistic", "jaccard", "text"),
  n = 100,
  col = segmenTools::arno(n),
  breaks,
  xpd = FALSE,
  show.total = FALSE,
  tot.cex = 0.8,
  file,
  ...
)
}
\description{
plots dotplot profiles of `clusterOverlaps' objects.
}
