% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{expandCircularFeatures}
\alias{expandCircularFeatures}
\title{Splits genome features crossing ends of circular chromosomes.
Splits genome features that cover start/end coordinates of circular
chromosomes, and adds the downstream half with optional modification
of ID, and type values. Note that only the upstream half retains
all column information, the downstream half will only carry
information on coordinates, and optionally updated feature type and ID.
The update will only happen if the passed table contains type and ID
information (see argument \code{idCools}. The split can be reversed
by function \code{removeCircularFeatures}.}
\usage{
expandCircularFeatures(features, chrL, coorCols = c("chr", "start", "end",
  "strand"), reverse = c("-", -1), idTag = "-circ2", idCols = c(ID = "ID",
  type = "type", parent = "parent"))
}
\arguments{
\item{features}{a list of genomic features with coordinates}

\item{chrL}{obligatory list of chromosome lengths, in order used
in chromosome column in \code{features} (see argument \code{coorCols}}

\item{coorCols}{ordered string vector providing the column names
of coordinate columns to be used; must be of length 4 and provide in
order: chromosome number (refering to argument \code{chrL}), start, end,
and strand (see argument \code{reverse})}

\item{reverse}{allowed indicators of reverse strand features
in strand column (see argument \code{coorCols})}

\item{idTag}{tag to add to downstream ID and type}

\item{idCols}{named vector of column names for feature ID, type,
and feature parent; note that a "parent" column will be added if not present
to refer the downstream half to its upstream feature, which retains
all other information}
}
\description{
Splits genome features crossing ends of circular chromosomes.
Splits genome features that cover start/end coordinates of circular
chromosomes, and adds the downstream half with optional modification
of ID, and type values. Note that only the upstream half retains
all column information, the downstream half will only carry
information on coordinates, and optionally updated feature type and ID.
The update will only happen if the passed table contains type and ID
information (see argument \code{idCools}. The split can be reversed
by function \code{removeCircularFeatures}.
}
\seealso{
\code{removeCircularFeatures}
}
