% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{figlabel}
\alias{figlabel}
\title{Add sub-figure label to corners of plots}
\usage{
figlabel(
  text,
  region = "figure",
  pos = "topleft",
  cex = NULL,
  shadow = FALSE,
  ...
)
}
\arguments{
\item{text}{figure label}

\item{region}{add to a figure, plot or device}

\item{pos}{position key, as in \code{legend}}

\item{cex}{label size}

\item{shadow}{use \code{\link{shadowtext}} to get background colors}

\item{...}{additional parameters for call to \code{text}}
}
\description{
copied from code by January Weiner 3rd at
\url{https://www.r-bloggers.com/adding-figure-labels-a-b-c-in-the-top-left-corner-of-the-plotting-region/},
originally published at
\url{https://logfc.wordpress.com/2017/03/15/adding-figure-labels-a-b-c-in-the-top-left-corner-of-the-plotting-region/}
}
