% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{fillGenome}
\alias{fillGenome}
\title{Get inter-segments}
\usage{
fillGenome(seg, chrS, indexed = TRUE, expand = TRUE)

fillGenome(seg, chrS, indexed = TRUE, expand = TRUE)
}
\arguments{
\item{seg}{input segments with indexed coordinates (see
\code{\link{coor2index}})}

\item{chrS}{a chromosome index, indicating at wich positions
chromosomes start; this is required for handling chromosome ends
and forward and reverse strand values}

\item{indexed}{boolean value indicating whether the coordinates
in argument \code{seg} have already been indexed; if not
\code{\link{coor2index}}) will be applied and the returned inter-segments
will be mapped back to chromosome coordinates by \code{\link{index2coor}}}

\item{expand}{if \code{TRUE} the returned inter-segments will contain
the same columns as the input; otherwise only coordinates are returned}

\item{seg}{input segments with indexed coordinates (see
\code{\link{coor2index}})}

\item{chrS}{a chromosome index, indicating at wich positions
chromosomes start; this is required for handling chromosome ends
and forward and reverse strand values}

\item{indexed}{boolean value indicating whether the coordinates
in argument \code{seg} have already been indexed; if not
\code{\link{coor2index}}) will be applied and the returned inter-segments
will be mapped back to chromosome coordinates by \code{\link{index2coor}}}

\item{expand}{if \code{TRUE} the returned inter-segments will contain
the same columns as the input; otherwise only coordinates are returned}
}
\description{
finds all ranges not present in the input segments, e.g. non-coding
regions in a list of coding regions

finds all ranges not present in the input segments, e.g. non-coding
regions in a list of coding regions
}
