% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTools.R
\name{findAACodon}
\alias{findAACodon}
\title{find genomic coordinates of amino acid position in a protein on genome
TODO: this does not account for introns}
\usage{
findAACodon(start, aa, strand)
}
\arguments{
\item{start}{start position of the ATG start codon}

\item{aa}{position of the amino acid in the protein}

\item{strand}{coding strand}
}
\description{
TODO: to solve this generally, the amino acid sequence of the protein
must be blasted against the genome
}
