% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{gamma.cdf}
\alias{gamma.cdf}
\title{calculates gamma distribution for given parameters}
\usage{
\method{gamma}{cdf}(x, start)
}
\arguments{
\item{x}{vector of x values for which the gamma distribution will
be calculated}

\item{start}{list containing gamma distribution parameters:
\code{a} (shape) and \code{b} (rate=1/scale)}
}
\description{
calculates gamma distribution for given parameters
}

