% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{get.fft}
\alias{get.fft}
\title{perform Discrete Fourier Transformation using \code{mvfft} from
the \code{stats} package, and returning the non-redundant (for
real numbers) first half of the transform, i.e., from the DC
(direct current) component to the Nyquist frequency}
\usage{
get.fft(x)
}
\arguments{
\item{x}{data to be transformed}
}
\description{
perform Discrete Fourier Transformation using \code{mvfft} from
the \code{stats} package, and returning the non-redundant (for
real numbers) first half of the transform, i.e., from the DC
(direct current) component to the Nyquist frequency
}

