% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{getChrS}
\alias{getChrS}
\title{generate chromosome index \code{chrS} from and ordered (!)
chromosome length table}
\usage{
getChrS(chrIdx, lcol = "length")
}
\arguments{
\item{chrIdx}{an ordered chromosome length table; feature
files for these chromosomes must have a column "chr" which
used the row number in chrIdx as its chromosome identifier}

\item{lcol}{numeric or character indicating the  number
or name of the column with chromosome lengths}
}
\description{
generate chromosome index \code{chrS} from and ordered (!)
chromosome length table
}
