% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{getClsAvg}
\alias{getClsAvg}
\title{calculates cluster averages
calculates average values and distributions for each cluster
and time point of a time series}
\usage{
getClsAvg(ts, cls, cls.srt, avg = get("median", mode = "function"),
  dev = get("sd", mode = "function"), q = 0.1)
}
\arguments{
\item{ts}{a time series, with time points in columns}

\item{cls}{a clustering of the time series, \code{length(cls)}
must equal \code{nrow(cls)}}

\item{cls.srt}{optional sorting of the clusters}

\item{avg}{a function for calculating an `average' value for
each cluster; default is the \code{median}}

\item{dev}{a function for calculating deviation from the
the average, default is the standard deviation (\code{sd})}

\item{q}{the \code{\link{quantile}} cut-off, a real number between 0 and 1;
\code{1-q} of the time-series are smaller and \code{q}; the range
of reported quantile time-series `high' and `low' will cover
(1- 2q) of all values in the cluster
of the time series are larger then the reported values.}
}
\description{
calculates cluster averages
calculates average values and distributions for each cluster
and time point of a time series
}
