% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{getSegmentClassTable}
\alias{getSegmentClassTable}
\title{splits segmenTier classes into a table}
\usage{
getSegmentClassTable(sgtypes, sep = "_", gsep = ":")

getSegmentClassTable(sgtypes, sep = "_", gsep = ":")
}
\arguments{
\item{sgtypes}{a list of strings, that is converted to a table
of classes based on a string separator (\code{sep})}

\item{sep}{the separator for classes in the string}

\item{gsep}{a separator within classes
to be used as classification ID (column header of the produces
class table)}

\item{sgtypes}{a list of strings, that is converted to a table
of classes based on a string separator (\code{sep})}

\item{sep}{the separator for classes in the string}

\item{gsep}{a separator within classes
to be used as classification ID (column header of the produces
class table)}
}
\description{
specifically tailored to segment strings in segmenTier;
splits a list of strings by a separator and constructs
a table from the entries.

specifically tailored to segment strings in segmenTier;
splits a list of strings by a separator and constructs
a table from the entries.
}
