% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gff2tab}
\alias{gff2tab}
\title{parse a GFF3 file into a table}
\usage{
gff2tab(file, attrsep = ";", fieldsep = "=")
}
\arguments{
\item{file}{gff file name}

\item{attrsep}{main separator of attribute string fields}

\item{fieldsep}{separator for attribute <id>=<value> couples}
}
\description{
Parses gff3 files into tables, including conversion of all attributes
}
