% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{idx2coor}
\alias{idx2coor}
\title{Simple version of \code{\link{index2coor}} for single values}
\usage{
idx2coor(pos, chrS, strands = c(1, -1))
}
\arguments{
\item{pos}{the continuous index position that will be mapped to
chromosome coordinates}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information}

\item{strands}{forward/reverse strand indicators}
}
\description{
Simple version of \code{\link{index2coor}} for single values
}

