% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{image_clustering}
\alias{image_clustering}
\title{plot sorted clustering as color table}
\usage{
image_clustering(cset, K, ...)
}
\arguments{
\item{cset}{a structure of class 'clustering' as returned by
segmenTier's \code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}}

\item{K}{cluster number; if missing the `selected' clustering will
be taken}

\item{...}{arguments to \code{\link{image_matrix}}}
}
\description{
plot sorted clustering as color table
}
