% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{image_matrix}
\alias{image_matrix}
\title{wrapper for \code{\link[graphics]{image}} plotting a data matrix
in the orientation of text display}
\usage{
image_matrix(
  z,
  x,
  y,
  text,
  text.col,
  text.cex = 1,
  axis = 1:2,
  axis.cex = 1.5,
  cut = FALSE,
  breaks,
  axis1.col,
  axis1.las = 2,
  axis2.col,
  axis2.las = 2,
  ...
)
}
\arguments{
\item{z}{the numeric data matrix to be plotted}

\item{x}{optional x coordinates, corresponding to columns of z}

\item{y}{optional y coordinates, corresponding to rows of z}

\item{text}{a matrix of characters corresponding to \code{dat}
which will be plotted on the image}

\item{text.col}{individual colors for text fields}

\item{text.cex}{relative font size for text fields}

\item{axis}{integer vector, sets whether x-axis (1,3) and/or
y-axis (2,4) are drawn; the column and row names of \code{dat} will
be used as tick labels}

\item{axis.cex}{if axis[1|2].col is provided, this sets the tick label size}

\item{cut}{if TRUE data is cut at min and max of breaks.}

\item{breaks}{breakst to use for color selection.}

\item{axis1.col}{invididual colors for x-axis tick labels, length must
equal the number of columns of \code{dat}}

\item{axis1.las}{parameter \code{las} (tick label orientation) for x axis}

\item{axis2.col}{invididual colors for y-axis tick labels, length must
equal the number of rows of \code{dat}}

\item{axis2.las}{parameter \code{las} (tick label orientation) for y axis}

\item{...}{further arguments to \code{\link[graphics]{image}}, e.g., col
to select colors}
}
\description{
Wrapper around \code{\link[graphics]{image}} to plot a matrix as
it is displayed in R console, i.e. the field \code{dat[1,1]}
is at the top left corner. It further allows to plot text
into individual fields and have colored axis tick labels.
}
