% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{lg2r}
\alias{lg2r}
\title{log2 ratio normalization}
\usage{
lg2r(x, na.rm = TRUE)
}
\arguments{
\item{x}{data to be transformed}

\item{na.rm}{remove NA values for mean calculation}
}
\value{
log2(x/apply(x,1,mean))
}
\description{
log2 ratio normalization
}
