% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{linreg}
\alias{linreg}
\title{Linear regression and correlation test, reporting
alpha, beta, the correlation and a p-value}
\usage{
linreg(x, Y, method = "pearson")
}
\arguments{
\item{x}{x vector}

\item{Y}{matrix of y vectors, with ncol(y)==length(x).}
}
\description{
Linear regression and correlation test, reporting
alpha, beta, the correlation and a p-value
}
