% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2df}
\alias{list2df}
\title{converts a names list of vectors of different length to a matrix
(data.frame) with names (now duplicated) and list elements in
columns 1 and 2.}
\usage{
list2df(x)
}
\arguments{
\item{x}{a named list of vectors}
}
\description{
converts a names list of vectors of different length to a matrix
(data.frame) with names (now duplicated) and list elements in
columns 1 and 2.
}
