% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{ma}
\alias{ma}
\title{moving average using \code{\link[stats]{filter}}}
\usage{
ma(x, n = 5, circular = FALSE)
}
\arguments{
\item{x}{data vector along which a moving average will be calculated}

\item{n}{moving average window size (integer!)}

\item{circular}{logical see help of function \code{\link[stats]{filter}}}
}
\description{
moving average using \code{\link[stats]{filter}}
}
