% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{meanzero}
\alias{meanzero}
\title{mean-0 normalization}
\usage{
meanzero(x)

meanzero(x)
}
\arguments{
\item{x}{data to be transformed}

\item{x}{data to be transformed}
}
\value{
t(apply(x,1,scale))

t(apply(x,1,scale))
}
\description{
mean-0 normalization

mean-0 normalization
}
