% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTools.R
\name{mutatePositions}
\alias{mutatePositions}
\title{mutate positions in a string}
\usage{
mutatePositions(seq, mutations)
}
\arguments{
\item{seq}{a string to be mutated}

\item{mutations}{a list of strings of the format 'A3T'
where the A at position 3 of the string should be changed
to a T}
}
\description{
mutate positions in a string
}
