% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osciTools.R
\name{normalize.los}
\alias{normalize.los}
\title{LVS/LOS normalization}
\usage{
normalize.los(x, los.id, norm.ref = c("median", "mean"),
  norm.type = c("loess", "smooth", "scale"), add.maximum = FALSE,
  reset.extrema = FALSE, extend.max = FALSE)
}
\arguments{
\item{x}{a data matrix to be normalized}

\item{los.id}{a numerical or logical vector, providing the indices
or a TRUE/FALSE vector for a set of "least-oscillating" or "least-variant"
data rows}

\item{norm.ref}{take the "median" or the "mean" of the reference
data set (\code{los.id}) for normalization}

\item{norm.type}{use "loess", "smooth" or "scale" normalization; see Details}

\item{add.maximum}{untested, see Details}

\item{reset.extrema}{untested, see Details}

\item{extend.max}{untested, see Details}
}
\description{
This function normalizes the columns of a matrix by a set of
invariant data rows (\code{LOS}, a "least-oscillating set") to be
defined by the user (option \code{los.id}). For each column a LOESS
fit, local fitting of a 2nd order polynomial with default
parameters of the R \code{stats} function
\code{\link[stats:loess]{loess}} is calculated between the values
of the reference set (\code{LOS}) and their median or median
(option \code{norm.ref} over all columns.  This is a modified
version of the function `normalize_LVS` from package
\code{FLUSH.LVS} (Calza et al. 2008), as used for normalization by
a "least-oscillating" set of "microarray" probes from a
transcriptome time-series of "yeast respiratory oscillations" by
Machne & Murray, 2012. Alternatively to LOESS but un-tested, the 
data can also be normalized by smoothing or simple min/max scaling.
Also un-tested and highly experimental: to avoid cutting data at
highest values of the \code{LOS} set, one can \code{add.maxima} of each
column to the LOESS fits, or after the fit and normalization reset
(\code{reset.extrema}) or re-scale (\code{extend.max}) maximal values.
}
\references{
Calza S, Valentini D, Pawitan Y (2008)
Normalization of oligonucleotide arrays based on the least-variant
set of genes. BMC Bioinformatics 9:140.

Machne R, Murray DB (2012)
The yin and yang of yeast transcription: elements of a global feedback
system between metabolism and chromatin. PLoS One 7(6):e37906
}
