% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{parseAnnotationList}
\alias{parseAnnotationList}
\title{Parse a matrix of ID/annotation mappings}
\usage{
parseAnnotationList(got, sep = ";")
}
\arguments{
\item{got}{a 2D matrix with IDs in the first column and a list of
annotation terms in the second}

\item{sep}{separator used for the list of terms in the second
column}
}
\description{
Parses a 2D ID/annotation mapping where annotations are a list of
terms with a separator. The function returns a TRUE/FALSE table that
can be used as input to \code{\link{clusterAnnotation}}.
}
