% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{parseGO}
\alias{parseGO}
\title{Generate a gene ontology (GO) table from a table containing
a column with gene IDs and a column with a list of GO terms.}
\usage{
parseGO(
  genes,
  obo.file,
  cols = c("ID", "GO"),
  sep = ",",
  short.names = TRUE,
  add.roots = TRUE,
  rm.obsolete = FALSE,
  minimal.set = FALSE
)
}
\arguments{
\item{genes}{a table of gene annotations containing two columns, one providing gene IDs and another providing a list of GO terms.}

\item{cols}{column names with gene IDs and GO term lists.}

\item{sep}{separator used in the GO term list column.}

\item{add.roots}{find roots and add to result.}

\item{rm.obsolete}{remove terms tagged as obsolete in the
desription.}

\item{minimal.set}{reduce GO terms to a minimal set of
non-redundant terms via
\link[ontologyIndex:minimal_set]{ontologyIndex::minimal_set}.}
}
\description{
Generate a gene ontology (GO) table from a table containing
a column with gene IDs and a column with a list of GO terms.
}
