% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{parseOBO}
\alias{parseOBO}
\title{Get GO (gene ontology) term descriptions from the ontology
definition file (.obo), using the ontologyIndex library.}
\usage{
parseOBO(
  file,
  terms,
  add.roots = FALSE,
  rm.obsolete = FALSE,
  minimal.set = FALSE
)
}
\arguments{
\item{file}{gene ontology definition file (.obo).}

\item{terms}{GO terms for which descriptions should be searched.}

\item{add.roots}{find roots and add to result.}

\item{rm.obsolete}{remove terms tagged as obsolete in the
desription.}

\item{minimal.set}{reduce GO terms to a minimal set of
non-redundant terms via
\link[ontologyIndex:minimal_set]{ontologyIndex::minimal_set}.}
}
\description{
Get GO (gene ontology) term descriptions from the ontology
definition file (.obo), using the ontologyIndex library.
}
