% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{phaseDist}
\alias{phaseDist}
\title{Calculates phase distributions}
\usage{
phaseDist(phs, w, degrees = TRUE)
}
\arguments{
\item{phs}{vector of phases in degrees}

\item{w}{optional weights for weighted mean}

\item{degrees}{phases are in degrees (0-360), set FALSE for radians}
}
\description{
calculates the circular mean and R statistics, copied from package
\code{CircStats} and after
http://en.wikipedia.org/wiki/Directional_statistics#Measures_of_location_and_spread and
https://en.wikipedia.org/wiki/Mean_of_circular_quantities
TODO: recognize bimodal?
}
