% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{plot.cdfLst}
\alias{plot.cdfLst}
\title{Some plotting routines for overlap statistics from
\code{\link{getOverlapStats}}}
\usage{
\method{plot}{cdfLst}(x = seq(0, 2, 0.05), CDF, type = "rcdf", col, lty,
  h = c(0.2, 0.8), v = c(0.8, 1.2), ylab = "cum.dist.func.", ...)
}
\description{
Some plotting routines for overlap statistics from
\code{\link{getOverlapStats}}
}

