% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plot.clusteraverages}
\alias{plot.clusteraverages}
\title{plots cluster averages}
\usage{
\method{plot}{clusteraverages}(x, cls.srt, cls.col, each = FALSE,
  polygon = TRUE, xlab, time, ylab = "average", ylim = ifelse(each, "avg",
  "rng"), ylim.scale = 0.1, ...)
}
\arguments{
\item{x}{cluster time series average object as calculated by
\code{\link{clusterAverages}}}

\item{cls.srt}{optional sorting of clusters; clusters will be plotted
in this order, i.e. the last in \code{cls.srt} is plotted last}

\item{cls.col}{optional coloring of clusters}

\item{each}{logical value indicating whether to plot all cluster
averages on one panel (\code{FALSE}) or each cluster on a separate panel
(\code{TRUE})}

\item{polygon}{logical indicating whether to plot the ranges `high' and
`low' in the cluster average object \code{avg}}

\item{xlab}{x-axis label}

\item{time}{optional time-points of the x-axis}

\item{ylab}{y-axis label}

\item{ylim}{either conventional range of the y-axis, or a string
specifying whether ylim should be calculated from the average
(\code{ylim="avg"}) or from the lower/upper ranges (\code{ylim="rng"})}

\item{ylim.scale}{if \code{ylim=="avg"}, the calculated ylim will be
extended by this fraction of the total range on both sides}

\item{...}{arguments to plot}
}
\description{
plots average time series of clusters as calculated by
\code{\link{clusterAverages}}, including the variations around the mean
as polygons
}
