% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotBIC}
\alias{plotBIC}
\title{plot BIC from flowclusterTimeseries}
\usage{
plotBIC(cls)
}
\arguments{
\item{cls}{"clustering" object from function
\code{\link[segmenTier:flowclusterTimeseries]{flowclusterTimeseries}}}
}
\description{
plot BIC from flowclusterTimeseries
}
