% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotClsAvg}
\alias{plotClsAvg}
\title{plots cluster averages
plots average time series of clusters as calculated by
\code{\link{getClsAvg}}, including the variations around the mean
as polygons}
\usage{
plotClsAvg(avg, cls.srt, cls.col, each = FALSE, polygon = TRUE, xlab, time,
  ylab = "average", ylim, ylim.scale = 0.1)
}
\arguments{
\item{avg}{cluster time series average object as calculated by
\code{\link{getClsAvg}}}

\item{cls.srt}{optional sorting of clusters; clusters will be plotted
in this order, i.e. the last in \code{cls.srt} is plotted last}

\item{cls.col}{optional coloring of clusters}

\item{each}{logical value indicating whether to plot all cluster
averages on one panel (\code{FALSE}) or each cluster on a separate panel
(\code{TRUE})}

\item{polygon}{logical indicating whether to plot the ranges `high' and
`low' in the cluster average object \code{avg}}

\item{xlab}{x-axis label}

\item{time}{optional time-points of the x-axis}

\item{ylab}{y-axis label}

\item{ylim}{range of the y-axis, will be calculated from the average
values, extended \code{ylim.scale}}

\item{ylim.scale}{if ylim is missing, the calculated ylim will be
extended by this fraction of the total range on both sides}
}
\description{
plots cluster averages
plots average time series of clusters as calculated by
\code{\link{getClsAvg}}, including the variations around the mean
as polygons
}
