% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotCor}
\alias{plotCor}
\title{scatter plot with statistics}
\usage{
plotCor(
  x,
  y,
  cor.method = c("pearson", "kendall", "spearman"),
  line.methods = c("ols", "tls"),
  line.col = c(1, 2),
  line.type = c(1, 2),
  signif = 1,
  round = 1,
  density = TRUE,
  col,
  ...
)
}
\arguments{
\item{x}{the x coordinates of the points in the plot.}

\item{y}{the y coordinates of the points in the plot.}

\item{cor.method}{method to calculate correlation and p-value via
\code{\link[stats:cor.test]{cor.test}}.}

\item{line.methods}{regression line methods, where \code{"ols"} is
a linear regression (ordinary least squares) with R's
\code{\link[stats:lm]{lm}} and \code{"tls"} (total least squares)
is calculated wih \code{\link[tls:tls]{tls}}.}

\item{line.col}{colors of the regression lines drawn with
\code{line.methods}.}

\item{line.type}{line types of the regression lines drawn with
\code{line.methods}.}

\item{signif}{number of digits to be shown for p-values in the plot
legend (argument \code{digits} to \code{\link{signif}}.}

\item{round}{number of decimal places to be shown for correlation
values in the plot legend (argument \code{digits} to
\code{\link{round}}.}

\item{density}{indicate local point densities by using
\link{dense2d} instead of the base \code{\link{plot}} function.}

\item{col}{color(s) of plotted points, if \code{density=FALSE}, or
color palette function if \code{density=TRUE} (argument
\code{colf} to \code{\link{dense2d}}).}

\item{...}{further arguments to the plotting function;
\code{\link{plot}} or, if \code{density=TRUE},
\code{\link{dense2d}}.}
}
\description{
scatter plot with statistics
}
