% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotCorMulti}
\alias{plotCorMulti}
\title{Plot linear regressions for subsets (clusters) of data}
\usage{
plotCorMulti(
  x,
  y,
  col,
  cls,
  cls.col,
  cls.srt,
  alpha = 0.5,
  method = "pearson",
  line = "ols",
  leg.pos = "bottomright",
  ...
)
}
\arguments{
\item{x}{the x coordinates of the points in the plot.}

\item{y}{the y coordinates of the points in the plot.}

\item{col}{point colors; if not provided cls.col will be used.}

\item{cls}{a vector of cluster labels for the xy data.}

\item{cls.col}{a named vector of cluster label colors.}

\item{cls.srt}{a sorting and sub-selection of cluster labels.}

\item{alpha}{color alpha level applied to cluster colors of the
plotted points.}

\item{method}{method used to calculate the correlations.}

\item{...}{arguments to \link{plotCor}.}
}
\description{
Plot linear regressions for subsets (clusters) of data
}
