% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotDFT}
\alias{plotDFT}
\title{plot polar coordinates}
\usage{
plotDFT(dft, col, cycles = 3, radius = 0.9, lambda = 1, bc = "component", ...)
}
\arguments{
\item{dft}{the Fourier transform of a time series as returned
by \code{t(mvfft(t(timeseries)))}, or alternatively, a `timeseries' object
from segmenTier's
\code{\link[segmenTier:processTimeseries]{processTimeseries}} when
run with (\code{use.fft=TRUE})}

\item{col}{a color vector for the rows in argument \code{dft} or
alternatively,  `clustering' object as returned by
segmenTier's \code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}
with coloring information}

\item{cycles}{the number of cycles (index of non-DC DFT component)
to be plotted}

\item{radius}{radius of the polar plot circle as a fraction
of data to be contained within the radius (smaller amplitude)}

\item{lambda}{parameter for Box-Cox transformation of DFT data; has no
effect for \code{lambda==1}}

\item{bc}{type of Box-Cox transformation (\code{if lambda!=1});
"component": separate transformation of real and imaginary parts of
the DFT; "amplitude": Box-Cox transformation of the amplitude}

\item{...}{arguments to the base \code{\link[graphics:plot]{plot}} 
and/or \code{\link[graphics:points]{points}} functions}
}
\description{
Plots the components of a Discrete Fourier Transform (DFT)
as polar coordinates (Re and Im of the complex numbers in the DFT).
Arguments \code{dft} and \code{col} can be segmenTier timeseries
and clustering objects.
}
