% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotOverlapsLegend}
\alias{plotOverlapsLegend}
\title{plot a legend for \code{\link{plotOverlaps}} plots}
\usage{
plotOverlapsLegend(
  p.min = 1e-10,
  p.txt = 1e-05,
  type = 1,
  round = 0,
  show.text = TRUE,
  l = 5,
  n = 100,
  col,
  dir = 1,
  labels = TRUE,
  ...
)
}
\arguments{
\item{p.min}{significance cutoff, p-values equal or smaller to
this cutoff will appear black (one-sided tests) or red/blue
(two-sided tests)}

\item{p.txt}{p-value cutoff for showing overlap numbers as white instead
of black text}

\item{type}{1-sided or 2-sided tests}

\item{round}{round parameter for numeric text}

\item{show.text}{show log10(p) as text fields (to indicated text p-value
cutoff)}

\item{l}{number of fields to show in plot}

\item{n}{number of color shades between \code{p=1} (white)
and \code{p >= p.min} (black)}

\item{col}{color ramp, default are grey values (one-sided) or
red (above) and blue (below) for two-sided tests, length of
this vector overrules parameter \code{n}}

\item{dir}{horizontal (1) or vertical (2) orientation}

\item{labels}{add axis labels}

\item{...}{further arguments to \code{\link{plotOverlaps}}}
}
\description{
plot a legend for \code{\link{plotOverlaps}} plots
}
