% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotSingles}
\alias{plotSingles}
\title{plot indivividual time-courses}
\usage{
plotSingles(x, cls, goi, each, lwd = 2, leg.xy = "topleft", ...)
}
\arguments{
\item{cls}{"clustering" object (simple list)}

\item{tset}{"timeseries" object from function
\code{\link[segmenTier:processTimeseries]{\processTimeseries}}}
}
\description{
plot indivividual time-courses (GOI: genes of interest) from
timeseries and clustering object (simple list)
}
