% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{pruneSegments}
\alias{pruneSegments}
\title{prune segments at chromosome ends}
\usage{
pruneSegments(
  x,
  chrL,
  chr = "chr",
  coors = c(start = "start", end = "end", coor = "coor"),
  remove.empty = FALSE,
  verb = 1
)
}
\arguments{
\item{x}{genome coordinate table,}

\item{chrL}{chromosome length vector, where the chromosome column
must provide the index in this vector; e.g. for 16 chromosomes,
chrL has length 16 and provides, in this order, the lengths of
chromosomes 1 to 16),}

\item{chr}{chromosome column name,}

\item{coors}{coordinate column names to cut,}

\item{remove.empty}{after pruning, remove any segments with length 0;
NOTE: that this also affects segments that were not pruned, and already
had length 0 in \code{x}.}

\item{verb}{integer level of verbosity, 0: no messages, 1: show messages}
}
\description{
Cuts segments coordinates that are beyond chromosome coordinates.
}
