% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{randomSegments}
\alias{randomSegments}
\title{randomize locations of input segments}
\usage{
randomSegments(query, qclass, total)
}
\arguments{
\item{query}{query set of segments to be randomized}

\item{qclass}{column name which holds a sub-classification (clustering) of
the query segments, omit or pass empty string ("") to use all}

\item{total}{total length of the query range (genome length), if missing
the start of the first segment is also used as end.}
}
\description{
randomizes the locations of input segments, while maintaining
the length distributions, optionally different distributions
of different segment classes. Coordinates have been converted to a
continuous index over all chromosomes with \code{\link{coor2index}}.
Note, that this ignores chromosome borders!
}
\details{
The function is also used for permutation tests in
\code{\link{segmentJaccard}}.
}
