% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTools.R
\name{range2nt}
\alias{range2nt}
\title{calculates letter frequences in ranges of strings}
\usage{
range2nt(seq, ranges, coors = c(chr = "chr", start = "start", end = "end"))
}
\arguments{
\item{seq}{a sequence object as returned by \code{readFasta}}

\item{ranges}{a table provided ranges in each string}

\item{coors}{coordinate names in ranges table}
}
\description{
calculates letter frequences in ranges of strings, e.g.,
nucleotide sequences. The input is a sequence object
as returned by (genomeBrowser's version of) \code{readFasta}.
TODO: move readFasta here and document!
}
