% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{readDist}
\alias{readDist}
\title{distribution of read-counts, i.e., the mean, var, min&max and the number
of NA values}
\usage{
readDist(rds)

readDist(rds)
}
\arguments{
\item{rds}{vector of read-counts, e.g., within a genomic interval (segment)}

\item{rds}{vector of read-counts, e.g., within a genomic interval (segment)}
}
\description{
distribution of read-counts, i.e., the mean, var, min&max and the number
of NA values

distribution of read-counts, i.e., the mean, var, min&max and the number
of NA values
}
