% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{relabelClusters}
\alias{relabelClusters}
\title{relabels cluster labels by their sorting}
\usage{
relabelClusters(cls)
}
\arguments{
\item{cls}{the `clustering' object from segmenTier's 
\code{\link[segmenTier:flowclusterTimeseries]{flowclusterTimeseries}}}
}
\description{
relabels all cluster labels in a `clustering' object (as returned
by \code{\link{clusterTimeseries2}},
\code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}} and
\code{\link[segmenTier:flowclusterTimeseries]{flowclusterTimeseries}})
by the cluster sorting (in \code{cls$sorting}), such that the new
numeric cluster labels reflect this sorting.
}
