% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTools.R
\name{revcomp}
\alias{revcomp}
\title{calculate reverse complement of RNA or DNA}
\usage{
revcomp(sq, type = "DNA", na = "N", reverse = TRUE)
}
\arguments{
\item{sq}{a, RNA or DNA string}

\item{type}{"RNA" or "DNA"}

\item{na}{string to use for non-recognized nucleotides}

\item{reverse}{revert to sequence}
}
\description{
Returns the reverse complement of RNA/DNA bases in an input string.
If \code{reverse==FALSE} it just reports complementary bases in the
same order.
}
