% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{runGost}
\alias{runGost}
\title{a wrapper for  \code{gprofiler2}'s \code{gost} function}
\usage{
runGost(
  cls,
  organism = "hsapiens",
  cls.srt,
  categories,
  terms,
  verb = 1,
  significant = FALSE,
  ...
)
}
\arguments{
\item{cls}{a name vector of a gene classification, where the names
are gene IDs recognized by \code{gprofiler2} in the respective
\code{organism},}

\item{organism}{the organism ID to which gene IDs of the \code{cls}
argument refer to; see \code{?gprofiler2::gost},}

\item{cls.srt}{an optional sorting of gene classes in \code{cls},}

\item{categories}{annotation source categories to report,}

\item{terms}{list of annotation terms to report (over all categories),
a name list of terms to report,}

\item{verb}{level of verbosity, 0: no output, 1: progress messages,}

\item{significant}{only report significant hits, option to \code{gost}.}

\item{...}{further arguments to \code{gprofiler2}'s \code{gost} function.}
}
\description{
takes a gene clustering, a named vector of gene categories, where
the names are gene IDs, and calls the \code{gost} function of the
\code{gprofiler2} to do enrichment analysis for all gene
annotations available. It then parses the output of the \code{gost}
function of the \code{gprofiler2} R package into overlap tables for
use with \code{segmenTools}' cluster analysis pipeline.
}
