% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{segmentAverage}
\alias{segmentAverage}
\title{Average read-counts of segments}
\usage{
segmentAverage(rds, avg = "mean", mean.ratio = FALSE,
  rm.extreme = FALSE, endcut = 0, k = 1, endrule = "median")
}
\arguments{
\item{rds}{matrix of read-count time-series}

\item{avg}{function to use as average statistics, \code{mean} or
\code{median}}

\item{mean.ratio}{calculate the mean ratio over time of each read-count
time-series before taking the total average}

\item{rm.extreme}{remove extreme value reads,
i.e., where the temporal mean is \code{<0.1} or \code{>0.9} of all means}

\item{endcut}{fraction of the ends the genomic interval to remove before
taking the total average}

\item{k}{parameter \code{k} for the running median
function \code{\link[stats]{runmed}}; the running median will be calculated
only if \code{k>1}}

\item{endrule}{parameter  \code{endrule} for the running median
function \code{\link[stats]{runmed}}}
}
\description{
Average read-count time-series of a genomic interval (segment). Note
that the diverse attempts to smooth or filter read-counts before
taking the averages did not give good results, and we remained using
avg="mean" and no other transformations.
}
