% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{segmentJaccard}
\alias{segmentJaccard}
\title{Jaccard-Index overlap test for classes of segments (genomic intervals)}
\usage{
segmentJaccard(query, target, qclass, tclass, perm = 0, verb = 1)
}
\arguments{
\item{query}{query set of segments}

\item{target}{target set of segments}

\item{qclass}{column name which holds a sub-classification (clustering) of
the query segments}

\item{tclass}{column name which holds a sub-classification (clustering) of
the target segments}

\item{perm}{number of permutations to perform}

\item{verb}{integer level of verbosity, 0: no messages, 1: show messages}
}
\description{
calculates the Jaccard index, including a permutation test, between
different classes in a query and a target set of segments
(genomic intervals), where coordinates have been converted to a
continuous index over all chromosomes with \code{\link{coor2index}}.
}
\details{
A simple permutation is performed, sampling randomly from
all inter-segment distances and segment lengths, and ignoring optional
query sub-classifications (argument \code{qclass}). Note, that chromosome
ends are ignored. The function is conceptually similar to
\code{\link{clusterCluster}} and results of the permutaiton
test (argument \code{perm>0}) can be plotted directly with
\code{\link{plotOverlaps}}.
}
