% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{segmentSort}
\alias{segmentSort}
\title{sort genomic coordinates}
\usage{
segmentSort(x, by.strand = FALSE, reverse = c("-1", "-"))
}
\arguments{
\item{x}{genomic interval (segments) table.}

\item{by.strand}{split by strand.}

\item{reverse}{reverse strand identifiers, all others are
considered to stem from the forward strand.}
}
\description{
Sorts genomic coordinates (genomic intervals) along chromosomes
}
