% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{selectColors}
\alias{selectColors}
\title{Map a distribution to a color scheme}
\usage{
selectColors(x, mn, mx, q = 0.1, colf = grDevices::gray.colors,
  reverse = FALSE, n = 100, plot = TRUE, xlab = "score", xlim,
  heights = c(0.75, 0.25), mai = c(0.75, 0.75, 0.1, 0.1), ylim,
  legendf, ...)
}
\arguments{
\item{x}{data vector}

\item{mn}{minimal value to cut data at, if missing
\code{\link[stats:quantile]{quantile}} will
be used with parameter \code{q} or \code{q[1]}}

\item{mx}{maximal value to cut data at, if missing the quantile
\code{q[2]} or, if only one {q} is provided, \code{1-q[1]} will be used}

\item{q}{parameter \code{probs} for quantile selection of
mn/mx data cuts, here a single value or two values; see \code{mn}
and \code{mx} arguments}

\item{colf}{color map function used to create a color gradient,
eg. \code{\link[grDevices:gray.colors]{gray.colors}} or
\code{viridis}}

\item{reverse}{revert color order}

\item{n}{number of colors, first argument to \code{colf}}

\item{plot}{plot a legend, ie. a histogram of the full data and an
\code{image} of the color scheme below}

\item{xlab}{x-axis label to use in legend plot}

\item{xlim}{optional limits to x-axis}

\item{heights}{relative heights of histogram and image plots}

\item{mai}{\code{par("mai")} plot parameter for plot margins}

\item{ylim}{y-axis limits for histogram}

\item{legendf}{legend plot function, called after the upper distribution
plot, can be used to add a legend}

\item{...}{further arguments to \code{colf}, e.g. \code{start} and
\code{end} in \code{\link[grDevices:gray.colors]{gray.colors}}}
}
\description{
Returns a list containing colors and breaks, to
be used in \code{image} plots, and the cut data and colors
for each datum. Note that the function is specifically used
for color scheme selection in the \code{genomeBrowser} project.
}
