% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqTools.R
\name{seq2nt}
\alias{seq2nt}
\title{calculate local nucleotide content}
\usage{
seq2nt(
  seq,
  window = 100,
  step = 1,
  abc,
  skew = "",
  cum = FALSE,
  circular = FALSE,
  verb = FALSE
)
}
\arguments{
\item{seq}{a sequence object as returned by \code{readFasta}}

\item{window}{width of the moving window}

\item{step}{step size}

\item{abc}{sequence letters for which averages are calculated}

\item{skew}{calculate skew for letters 1 and 2 in argument \code{abc}
(e.g. \code{abc=c("G","C"), skew=(G-C)/(G+C)})}

\item{cum}{set TRUE to calculate cumulative skew after Grigoriev 1998 NAR}

\item{circular}{treat sequence as circular}

\item{verb}{print progress messages}
}
\description{
calculate local nucleotide content in moving windows
and returns genomeData matrix
TODO: move readFasta here and document!
}
