% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{shadowtext}
\alias{shadowtext}
\title{plot borders around text based on text colors}
\usage{
shadowtext(
  x,
  y = NULL,
  labels,
  col = "white",
  bg = "black",
  auto.bg = TRUE,
  bg.thresh = 0.75,
  r = 0.05,
  ...
)
}
\arguments{
\item{x}{x-coordinates of text labels}

\item{y}{y-coordinates of text labels}

\item{labels}{vector of text labels}

\item{col}{vector of text foreground colors}

\item{bg}{vector of text background colors}

\item{auto.bg}{auto-select background color black or white based
on a brightness color model}

\item{bg.thresh}{brightness threshold to switch from black to white
background color}

\item{r}{ratio of stringwidth and stringheight to use for background color}

\item{...}{further arguments to \code{\link{text}} in both foreground
and background calls}
}
\description{
plot borders around text based on text colors
}
