% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{splitsegs}
\alias{splitsegs}
\title{splits segments that span chromosome ends}
\usage{
splitsegs(segs, chrS, idcol, verb = 0)
}
\arguments{
\item{segs}{a matrix of segment start and end coordinates given
in columns named "start" and "end"}

\item{chrS}{a chromosome index, indicating at wich positions
chromosomes start; this is required for handling chromosome ends
and forward and reverse strand values}

\item{idcol}{column holding segment IDs; IDs of split segments will
receive the suffix "_2" for once copy#'}

\item{verb}{integer level of verbosity, 0: no messages, 1: show messages}
}
\description{
Finds segments that span chromosomes ends and splits those
in two segments on each covered chromosome. The input must
contain columns "start" and "end"; these will be modified for
chromosome-spanning segments. All other entries in the matrix
will be copied, unless an "idcol" is specified, which will receive the
suffix "_2" for one of two copies.
}
