% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{switchStrand}
\alias{switchStrand}
\title{switches the strand information (reverse<->forward) of genomic
features with continuously indexed chromosome coordinates}
\usage{
switchStrand(features, chrS, cols = c("start", "end", "coor"))
}
\arguments{
\item{features}{genomic features with continuously indexed
chromosome coordinates}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information}

\item{cols}{names of the columns holding the continuous index}
}
\description{
switches the strand information (reverse<->forward) of genomic
features with continuously indexed chromosome coordinates
}

