% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{t.clusterOverlaps}
\alias{t.clusterOverlaps}
\title{transpose cluster overlap object}
\usage{
\method{t}{clusterOverlaps}(x)
}
\arguments{
\item{x}{object of class `clusterOverlaps`}
}
\description{
TODO: test this systematically!
}
\seealso{
\code{\link{clusterCluster}},
    \code{\link{clusterProfile}}, \code{\link{clusterAnnotation}}
}
