% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tab2gff}
\alias{tab2gff}
\title{creates a gff3-like table}
\usage{
tab2gff(
  tab,
  columns = c(seqid = "chr", source = "source", type = "type", start = "start", end =
    "end", score = "score", strand = "strand", phase = "phase"),
  attributes = c(ID = "ID", Name = "name", Alias = "alias", Parent = "parent", color =
    "color"),
  notes,
  sep = "; ",
  source = c("chromosome", "plasmid")
)
}
\arguments{
\item{tab}{input table of genomic features with chromosome coordinate
information}

\item{columns}{a named string vector mapping from columns in
\code{tab} (values) to the first 7 columns required for
the gff3 file (names)}

\item{attributes}{a named string vector mapping  from columns in
\code{tab} (values) to values in the attribute list; the names of
the vector will become the key in the attribute list
("<key> = <value>") and attributes will be separated by
arugment \code{sep}}

\item{notes}{as \code{arguments} but output will be
"note=<key>=<value>}

\item{sep}{separator for attribute list}

\item{source}{identifier of main source feature (chromosome)}
}
\description{
Creates a table with columns as present in gff3 genome
annotation files, including ;-separated attribute lists.
Columns with RGB colors can be used to create \code{snapgene}-specific
note in attributes.
}
