% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{tagDuplicates}
\alias{tagDuplicates}
\title{Tag duplicate names by increasing numbers.}
\usage{
tagDuplicates(names)
}
\arguments{
\item{names}{a vector of characters}
}
\description{
Tag duplicate names by increasing numbers.
}
