% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{w.test}
\alias{w.test}
\title{Wilcoxon test wrapper that returns a normalized U-statistic}
\usage{
w.test(x, y, ...)
}
\arguments{
\item{x, y}{as x and y passed to
\link[stats:wilcox.test]{wilcox.test}}

\item{...}{further parameters passed on to
\link[stats:wilcox.test]{wilcox.test}}
}
\description{
This wrapper just calls \link[stats:wilcox.test]{wilcox.test} but
normalizes the test's U-statistic by the product of the lengths
(counts of non-NA values) of x and y such that values U>0 indicate
that x is generally larger than y and opposite for U<0.
}
