% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{whichSegment}
\alias{whichSegment}
\title{which segment covers a position}
\usage{
whichSegment(pos, seg)

whichSegment(pos, seg)
}
\arguments{
\item{pos}{chromosome position in continuous index}

\item{seg}{segment (genomic interval)}

\item{pos}{chromosome position in continuous index}

\item{seg}{segment (genomic interval)}
}
\description{
which segment covers a position

which segment covers a position
}
