% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{writeSegments}
\alias{writeSegments}
\title{write out segment data to individual files}
\usage{
writeSegments(data, segments, path, name = "segment")
}
\arguments{
\item{data}{a data matrix to which coordinates in \code{segments}
refer to}

\item{segments}{a matrix that must contain "start" and "end" (columns)
of segments; these coordinates will be extracted from \code{data}
and written to individual files, numbered by the row number in segments.}

\item{path}{optional output path where files will be written, if not supplied
files will end up in the current working directory (`getwd`)}

\item{name}{name used as prefix in file names}
}
\description{
writing out data for segments to files, used for writing the primary
segments by \code{\link{presegment}} to single files that are
then further processed by segmenTier. This avoids loading the
complete data set and allows parallel usage of \code{segmenTier}.
}
