% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{alignData}
\alias{alignData}
\title{align genome data at specified coordinates (e.g. TSS)}
\usage{
alignData(coors, data, dst = 500, chrS, coorCols = c(chr = "chr",
  position = "coor", strand = "strand"), reverse = c("-", -1))
}
\arguments{
\item{coors}{genome positions (position, chromosome, strand)
at which data will be aligned
(TODO: allow start/end coors and set NA if beyond)}

\item{data}{genome data to be aligned; NOTE, that currently this
is required to be fully expanded matrix covering each position
(TODO: allow non-expanded data)}

\item{dst}{upstream/downstream length to be aligned
(TODO: allow different upstream and downstream ranges)
(TODO: allow individual ranges)}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information, see function \code{\link{getChrSum}}}

\item{coorCols}{ordered string vector providing the column names
of coordinate columns to be used; must be of length 3 and provide in
order: chromosome number (refering to argument \code{chrS}), position 
and strand (see argument \code{reverse})}

\item{reverse}{a vector of possible reverse strand indicators, all other
values in the strand column will be taken as forward strand!}
}
\description{
align genome data at specified coordinates (e.g. TSS)
}
