% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{annotateTarget}
\alias{annotateTarget}
\title{annotate target segments by overlapping query segments}
\usage{
annotateTarget(
  query,
  target,
  qcol = colnames(query),
  tcol,
  prefix,
  details = FALSE,
  only.best = TRUE,
  sort = TRUE,
  minJaccard,
  collapse = TRUE,
  msgfile = stdout()
)
}
\arguments{
\item{query}{the query set of segments (genomic intervals)}

\item{target}{the target set of segments (genomic intervals)}

\item{qcol}{column names to copy from target to it's best match in query,
defaults to all columns}

\item{tcol}{column names of target to include in the returned table,
defaults to none}

\item{prefix}{column name prefix for the copied columns}

\item{details}{set to \code{TRUE} to add details of the used
match, i.e., union and intersect, and relative position of the query
to the matching targets}

\item{only.best}{only consider the top-ranking query hit}

\item{sort}{sort hits by rank}

\item{minJaccard}{minimal Jaccard index (intersect/union) threshold;
overlaps below this value will NOT be reported; NOTE that this will
set argument \code{details} to \code{TRUE}}

\item{collapse}{if \code{TRUE} multiple query hits are collapsed
into a single row, with ;-separated strings in the respective fields.}

\item{msgfile}{file pointer for progress messages and warnings, defaults to
stdout, useful when using in context of command line pipes}
}
\description{
wrapper around \code{\link{segmentOverlap}}, used to 
annotate the target set by a column in the query set. See
Details for required input formats.
}
\details{
Wrapper around  \code{\link{segmentOverlap}}, used to 
annotate the target set by a column in the query set. Note that
coordinates must already be `indexed', see \code{?\link{segmentOverlap}}
and \code{?\link{coor2index}}. The default options (\code{only.best=TRUE},
\code{collapse=TRUE} yield a result matrix with 1 row for each
\code{target} (\code{nrow(results)==nrow(target)}, annotated by
only the best ranking overlapping \code{query} segment(s), ie. the
the \code{query} segment(s) with the highest Jaccard index. If multiple
\code{query} segments have the same Jaccard index, they are collapsed
into ;-separated lists. If \code{collapse=FALSE}, each overlapping
\code{query} segment will have its own row, and the result matrix
is longer (\code{nrow(results)>=nrow(target)}). If \code{only.best=FALSE},
all overlapping \code{query} segments will be reported, and optionally
sorted by rank (\code{sort}).
Additionaly a minimal Jaccard filter can be applied
(option \code{minJaccard}), and only \code{query} segments with a
higher Jaccard index are reported.
}
