% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{annotateTarget}
\alias{annotateTarget}
\title{annotate target segments by overlapping query segments}
\usage{
annotateTarget(query, target, qcol = colnames(query), tcol, prefix,
  details = FALSE, only.best = TRUE, minJaccard, collapse = TRUE,
  msgfile = stdout())
}
\arguments{
\item{query}{the query set of segments (genomic intervals)}

\item{target}{the target set of segments (genomic intervals)}

\item{qcol}{column names to copy from target to it's best match in query,
defaults to all columns}

\item{tcol}{column names of target to include in the returned table,
defaults to none}

\item{prefix}{column name prefix for the copied columns}

\item{details}{set to \code{TRUE} to add details of the used
match, i.e., union and intersect, and relative position of the query
to the matching targets}

\item{only.best}{only consider the top-ranking query hit}

\item{minJaccard}{minimal Jaccard index (intersect/union) threshold;
overlaps below this value will NOT be reported; NOTE that this will
set argument \code{details} to \code{TRUE}}

\item{collapse}{if \code{TRUE} multiple query hits are collapsed
into a single row, with ;-separated strings in the respective fields.}

\item{msgfile}{file pointer for progress messages and warnings, defaults to
stdout, useful when using in context of command line pipes}
}
\description{
wrapper around \code{\link{segmentOverlap}}, used to 
annotate the target set by a column in the query set. Note that
coordinates must already be `indexed', see \code{?\link{segmentOverlap}}
and \code{?\link{coor2index}}.
}
