% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{ash}
\alias{ash}
\title{asinh trafo, an alternative to log transformation that has less
(compressing) effects on the extreme values (low and high values),
and naturally handles negative numbers and 0}
\usage{
ash(x)
}
\arguments{
\item{x}{data to be transformed}
}
\description{
asinh trafo, an alternative to log transformation that has less
(compressing) effects on the extreme values (low and high values),
and naturally handles negative numbers and 0
}

