% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osciTools.R
\name{circ.dens}
\alias{circ.dens}
\title{circular density}
\usage{
circ.dens(x, bw = 18, freq = FALSE, ...)
}
\arguments{
\item{x}{phases in degree and of class numeric, NA will be removed}

\item{bw}{bandwith parameter of
\code{\link[circular:density.circular]{density.circular}}}

\item{freq}{if TRUE densities \code{y} will be scaled by the total
number of measurement \code{N} as \code{N * y/sum(y)}; the resulting
density function will integrate to \code{N}}

\item{...}{further arguments to
\code{\link[circular:density.circular]{density.circular}}}
}
\description{
wrapper around the \code{\link[circular:density.circular]{density.circular}}
function from package \code{circular}, that maps back to numeric
types, to avoid automatic handling of the S3 class circular data
(eg. polar plots) and allows to scale densities by the absolute number;
see argument \code{freq}
}
