% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{clusterAnnotation}
\alias{clusterAnnotation}
\title{Clustering enrichment scan
Scans for overlap enrichments of a clustering in a matrix of
clusterings, potentially simply a TRUE/FALSE table, eg. indicating
annotation with a specific Gene Ontology or other term. It reports
tables of all overlap sizes and their enrichment p-values. The
function is a wrapper around \code{\link{clusterCluster}}, which
performs cumulative hypergeometric distribution tests between
two clusterings.}
\usage{
clusterAnnotation(cls, data, cls.srt, p.thresh = 0.01, verbose = TRUE)
}
\arguments{
\item{cls}{a character or numeric vector with the main clustering}

\item{data}{a string, numeric or logical matrix with other categorizations
of genes. Data rows must correspond to clustering in argument \code{cls}}

\item{cls.srt}{a numeric ar string vector indicating a cluster sorting,
allows also to analyze only a subset of clusters in argument \code{cls}}

\item{p.thresh}{p-value threshold for cluster summary table}

\item{verbose}{print progress messages}
}
\description{
Clustering enrichment scan
Scans for overlap enrichments of a clustering in a matrix of
clusterings, potentially simply a TRUE/FALSE table, eg. indicating
annotation with a specific Gene Ontology or other term. It reports
tables of all overlap sizes and their enrichment p-values. The
function is a wrapper around \code{\link{clusterCluster}}, which
performs cumulative hypergeometric distribution tests between
two clusterings.
}
