% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{clusterCluster}
\alias{clusterCluster}
\title{calculates overlaps between two clusterings}
\usage{
clusterCluster(cl1, cl2, na.string = "na", cl1.srt, cl2.srt,
  req.vals = c("greater"))
}
\arguments{
\item{cl1}{clustering 1}

\item{cl2}{clustering 2}

\item{na.string}{replace NA or empty strings by `<na.string>'}

\item{cl1.srt}{optional cluster sorting of clustering 1}

\item{cl2.srt}{optional cluster sorting of clustering 2}

\item{req.vals}{requested statistics; one of `greater', `less'
or `two-sided' to calculate enrichment, depletion or the more
more signficant of both, respectively}
}
\description{
calculates mutual overlaps between two clusterings of the same data set
using hypergeometric distribution statistics for significantly
enriched or deprived mutual overlaps
}
