% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{clusterCluster}
\alias{clusterCluster}
\title{mututal overlaps between two clusterings of the same data set;
calculates hypergeometric distribution statistics for significantly
enriched or deprived mutual overlaps}
\usage{
clusterCluster(cl1, cl2, bonferroni = FALSE, store.data = FALSE,
  req.vals = c("greater"), na.rm = FALSE, verbose = TRUE, pval = 0.01,
  plot = FALSE, xlab = "cluster 1", ylab = "cluster 2",
  na.string = "na", plot.type = "spine", plot.ppoor = FALSE,
  plot.prich = TRUE, name1, name2, cl1.max = 100, cl2.max = 100,
  cl1.sorting, cl2.sorting)
}
\arguments{
\item{cl1}{clustering 1}

\item{cl2}{clustering 2}

\item{bonferroni}{bonferroni correction is merely for signficant messages when option 'verbose' is TRUE}

\item{plot.type}{default is a spineplot, matrices from the returned results
can be plotted as heat maps , e.g. by plot.type=="p.value" or
plot.type=="percent"}
}
\description{
mututal overlaps between two clusterings of the same data set;
calculates hypergeometric distribution statistics for significantly
enriched or deprived mutual overlaps
}

