% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{dense2d}
\alias{dense2d}
\title{2D density heatmap plot}
\usage{
dense2d(x, y, pch = 20, nbin = c(128, 128), ...)
}
\arguments{
\item{x}{x-coordinates}

\item{y}{y-coordinates}

\item{pch}{\code{pch} argument to plot}

\item{nbin}{number of bins in each dimension, see argument \code{nbins}
\code{\link[grDevices:densCols]{densCols}} and \code{gridsize}
to \code{\link[KernSmooth:bkde2D]{bkde2D}}}

\item{...}{arguments to plot (hint:cex can be useful)}
}
\value{
the plotted data.frame, including local densities
}
\description{
copied from Josh O'Brien posted at
https://stackoverflow.com/questions/17093935/r-scatter-plot-symbol-color-represents-number-of-overlapping-points/17096661#17096661
}
