% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{dense2d}
\alias{dense2d}
\title{2D density heatmap plot}
\usage{
dense2d(
  x,
  y,
  pch = 20,
  nbin = c(128, 128),
  circular = FALSE,
  x.circular = FALSE,
  y.circular = FALSE,
  xlim,
  ylim,
  colf = viridis::viridis,
  ...
)
}
\arguments{
\item{x}{x-coordinates}

\item{y}{y-coordinates}

\item{pch}{\code{pch} argument to plot}

\item{nbin}{number of bins for both dimensions, can be a single.
number for both dimensions, or separate numbers, see argument
\code{nbins} to function
\code{\link[grDevices:densCols]{densCols}} and \code{gridsize}
to \code{\link[KernSmooth:bkde2D]{bkde2D}}.}

\item{circular}{treat x and y as circular coordinates in radian.}

\item{x.circular}{treat x as circular coordinates in radian.}

\item{y.circular}{treat y as circular coordinates in radian.}

\item{xlim, }{numeric vectors of length 2, giving the x 
coordinate ranges.}

\item{ylim}{as \code{xlim} for y.}

\item{colf}{color map function used to create a color gradient,
eg. \code{\link[grDevices:colorRampPalette]{colorRampPalette}}
or \code{viridis}.}

\item{...}{arguments to plot (hint:cex can be useful).}
}
\value{
the plotted data.frame, including local densities.
}
\description{
Uses base R's \code{\link[grDevices:densCols]{densCols}} to
calculate local densities at each point in a scatterplot, and then
replaces them by a colored scheme, copied from Josh O'Brien posted
at
https://stackoverflow.com/questions/17093935/r-scatter-plot-symbol-color-represents-number-of-overlapping-points/17096661#17096661
}
