% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{getChrS}
\alias{getChrS}
\title{Generate chromosome index \code{chrS} from lengths}
\usage{
getChrS(chrL)
}
\arguments{
\item{chrL}{an ordered vector of chromosome lengths; where the
order must correspond to chromosome numbering in feature tables
for which chrS is used}
}
\description{
Generate chromosome index \code{chrS} from lengths
}
