% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{image_clustering}
\alias{image_clustering}
\title{plot sorted clustering as color table}
\usage{
image_clustering(cset, k = selected(cset), ...)
}
\arguments{
\item{cset}{a structure of class 'clustering' as returned by
segmenTier's \code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}}

\item{k}{integer or string specifiying to clustering (K: cluster numbers)
to be used if cls is of class 'clustering'; if missing (default) the
`selected' clustering from \code{cls} is chosen}

\item{...}{arguments to \code{\link{image_matrix}}}
}
\description{
plot sorted clustering as color table
}
