% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{image_matrix}
\alias{image_matrix}
\title{data.matrix `as-is' wrapper for \code{\link[graphics]{image}}}
\usage{
image_matrix(dat, text, text.col, text.cex = 1, axis = 1:2, axis1.col,
  axis1.las = 2, axis2.col, axis2.las = 2, axis.cex = 1.5, ...)
}
\arguments{
\item{dat}{the numeric data matrix to be plotted}

\item{text}{a matrix of characters corresponding to \code{dat}
which will be plotted on the image}

\item{text.col}{individual colors for text fields}

\item{text.cex}{relative font size for text fields}

\item{axis}{integer vector, sets whether x-axis (1,3) and/or
y-axis (2,4) are drawn; the column and row names of \code{dat} will
be used as tick labels}

\item{axis1.col}{invididual colors for x-axis tick labels, length must
equal the number of columns of \code{dat}}

\item{axis1.las}{parameter \code{las} (tick label orientation) for x axis}

\item{axis2.col}{invididual colors for y-axis tick labels, length must
equal the number of rows of \code{dat}}

\item{axis2.las}{parameter \code{las} (tick label orientation) for y axis}

\item{axis.cex}{if axis[1|2].col is provided, this sets the tick label size}

\item{...}{further arguments to \code{\link[graphics]{image}}, e.g., col
to select colors}
}
\description{
Wrapper around \code{\link[graphics]{image}} to plot a matrix as
it is displayed in R console, i.e. the field \code{dat[1,1]}
is at the top left corner. It further allows to plot text
into individual fields and have colored axis tick labels.
}
