% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{image_matrix}
\alias{image_matrix}
\title{Wrapper around \code{\link[graphics]{image}} to plot a matrix as
it is displayed in R console, i.e. the field \code{dat[1,1]}
is at the top left corner. It further allows to plot text
into individual fields and have colored axis tick labels.}
\usage{
image_matrix(dat, text, text.col, axis = 1:2, axis1.col, axis2.col, ...)
}
\arguments{
\item{dat}{the numeric data matrix to be plotted}

\item{text}{a matrix of characteres corresponding to \code{dat}
which will be plotted on the image}

\item{axis}{integer vector, sets whether bottom (1) and/or left
(2) axis are draw; the column and row names of \code{dat} will
be used as tick labels}

\item{axis1.col}{invididual colors for x-axis tick labels, length must
equal the number of columns of \code{dat}}

\item{axis2.col}{invididual colors for y-axis tick labels, length must
equal the number of rows of \code{dat}}

\item{...}{further arguments to \code{\link[graphics]{image}}, e.g., col
to select colors}
}
\description{
Wrapper around \code{\link[graphics]{image}} to plot a matrix as
it is displayed in R console, i.e. the field \code{dat[1,1]}
is at the top left corner. It further allows to plot text
into individual fields and have colored axis tick labels.
}

