% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{insertRows}
\alias{insertRows}
\title{Util to insert multiple rows at specified positions
in a \code{data.frame}, expanding single-row code by user
Ari B. Friedman at
\url{https://stackoverflow.com/questions/11561856/add-new-row-to-dataframe-at-specific-row-index-not-appendedlooping through new rows}}
\usage{
insertRows(existingDF, newrows, r)
}
\arguments{
\item{existingDF}{existing \code{data.frame}}

\item{newrows}{rows to add to \code{existingDF}}

\item{r}{positions at which rows are to be inserted, \code{length(r)}
must equal \code{nrow(newrows)}, and \code{r<=nrow(existingDF)}}
}
\description{
Util to insert multiple rows at specified positions
in a \code{data.frame}, expanding single-row code by user
Ari B. Friedman at
\url{https://stackoverflow.com/questions/11561856/add-new-row-to-dataframe-at-specific-row-index-not-appendedlooping through new rows}
}
