% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{log_1}
\alias{log_1}
\title{log trafo handling zeros by adding 1}
\usage{
log_1(x)

log_1(x)
}
\arguments{
\item{x}{data to be transformed}

\item{x}{data to be transformed}
}
\value{
log(x+1)

log(x+1)
}
\description{
log trafo handling zeros by adding 1

log trafo handling zeros by adding 1
}
