% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parseGEOSoft}
\alias{parseGEOSoft}
\title{parse GO Soft archives}
\usage{
parseGEOSoft(file, idcol = "ORF", valcol = "VALUE", title = TRUE,
  desc = TRUE)
}
\arguments{
\item{file}{a GEO Soft archive file (eg. GSE18902_family.soft.gz)}

\item{idcol}{ID columns from the platform table; NOTE: that the first
column in this vector}

\item{valcol}{value column (currently only 1 is allowed) in the sample tables}

\item{title}{if TRUE a sample title will be retrieved}

\item{desc}{if TRUE, value descriptions in field "#<valcol>" will be retrieved}
}
\description{
parses GEO Soft microarray experiment archives into a
data matrix and an ID-probe mapping table
}
