% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{parseJaccard}
\alias{parseJaccard}
\title{parses overlap statitics produced with bedtools}
\usage{
parseJaccard(ovfile, prefix, qclass = "query", tclass = "target")
}
\arguments{
\item{ovfile}{output file produced by \code{segmentoverlaps.sh}.}

\item{prefix}{optional class name prefix that was required for
bed file definition (name, score columns).}

\item{qclass}{optional name for the query classes.}

\item{tclass}{optional name for the target classes.}
}
\description{
Parses a file that is produced by the script \code{segmentoverlaps.sh}
(based on USCS bedtools), into an \code{clusterOverlaps} class.
}
