% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{phaseDist}
\alias{phaseDist}
\title{Calculates phase distributions in segments, i.e., the circular mean and
standard deviation and the R statistics, copied from from package
\code{circstats} and after
http://en.wikipedia.org/wiki/Directional_statistics#Measures_of_location_and_spread and
https://en.wikipedia.org/wiki/Mean_of_circular_quantities
TODO: recognize bimodal?}
\usage{
phaseDist(phs, w)
}
\arguments{
\item{phs}{vector of phases in degrees}

\item{w:}{optional weights for weighted mean}
}
\description{
Calculates phase distributions in segments, i.e., the circular mean and
standard deviation and the R statistics, copied from from package
\code{circstats} and after
http://en.wikipedia.org/wiki/Directional_statistics#Measures_of_location_and_spread and
https://en.wikipedia.org/wiki/Mean_of_circular_quantities
TODO: recognize bimodal?
}

