% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotClusters}
\alias{plotClusters}
\title{plots cluster averages}
\usage{
plotClusters(x, cls, K, cls.col, cls.srt, each = TRUE, type = "rng", norm,
  avg = "median", q = 0.9, ylab, ylim = ifelse(each, "avg", "rng"),
  ylim.scale = 0.1, time, xlab, avg.col = "#000000", ...)
}
\arguments{
\item{x}{either a simple data matrix with time points (x-axis) in columns,
or a processed time-series as provided by
\code{\link[segmenTier:processTimeseries]{processTimeseries}}}

\item{cls}{eiter a vector (length(cls)==nrow(x) or a structure of
class 'clustering' as returned by segmenTier's
\code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}}

\item{K}{integer or string specifiying to clustering (K: cluster numbers)
to be used if cls is of class 'clustering'; if missing (default) the
`selected' clustering from \code{cls} (cls$selected) is chosen}

\item{cls.col}{optional named cluster color vector, where names indicate
the clusters (as.caracter(cls)); if cls is of class 'clustering' it
is taken from there}

\item{cls.srt}{optional cluster sorting, can be used for selection of
subsets of clusters; if cls is of class 'clustering' it is taken
from there}

\item{each}{logical value indicating whether to plot all cluster
averages on one panel (\code{FALSE}) or each cluster on a separate panel
(\code{TRUE})}

\item{type}{string specifying the type of plot: "rng" for plotting
only data ranges (see argument \code{q}) or "all" to plot
each individual time-course (as thin lines)}

\item{norm}{normalization of the time-series data, must be a function
that transforms the data, available via \link{segmenTools} are
\code{lg2r}, \code{ash}, \code{log_1}, \code{meanzero} normalizations}

\item{avg}{a function for calculating an `average' value for
each cluster; default is the \code{median}}

\item{q}{either numeric 0-1, the fraction of data to be shown,
or a function name for calculating variance (eg. "sd", "var")}

\item{ylab}{y-axis label (only used if \code{each==FALSE})}

\item{ylim}{either conventional range of the y-axis, or a string
specifying whether ylim should be calculated from the average
(\code{ylim="avg"}) or from the lower/upper ranges (\code{ylim="rng"})}

\item{ylim.scale}{if \code{ylim=="avg"}, the calculated ylim will be
extended by this fraction of the total range on both sides}

\item{time}{optional numeric vector specifiying x-axis time coordinates}

\item{xlab}{x-axis label (auto-selected if missing)}

\item{avg.col}{color for average line; used only if \code{rng="all"}}

\item{...}{further arguments to \code{\link{plot.clusteraverages}}}
}
\description{
as \code{\link{plot.clusteraverages}} but using
package \link[segmenTier]{segmenTier}'s time series and clustering object
as input; calls, and plots and returns results from
\code{\link{clusterAverages}}.
}
