% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotCor}
\alias{plotCor}
\title{Scatter plot with correlation statistics.}
\usage{
plotCor(
  x,
  y,
  outliers,
  classes,
  cor.method = c("pearson", "kendall", "spearman"),
  line.methods = c("ols", "tls"),
  log = "",
  na.rm = TRUE,
  verb = 0,
  circular = FALSE,
  circular.jitter = "",
  cor.legend = TRUE,
  line.legend = FALSE,
  title = FALSE,
  line.col = c(1, 2),
  pch = 20,
  cex = 1,
  axes = TRUE,
  legpos,
  legbg = "#FFFFFF99",
  legcex = 1,
  signif = 1,
  round = 2,
  density = TRUE,
  col,
  ...
)
}
\arguments{
\item{x}{the x coordinates of the points in the plot.}

\item{y}{the y coordinates of the points in the plot.}

\item{outliers}{vector of indices or logical vector of x,y values
to exclude from correlation analysis.}

\item{classes}{classification of all values (length(x)), for which
correlations are calculated separateley.}

\item{cor.method}{method to calculate correlation and p-value via
\code{\link[stats:cor.test]{cor.test}}.}

\item{line.methods}{regression line methods, where \code{"ols"} is
a linear regression (ordinary least squares) with R's
\code{\link[stats:lm]{lm}} and \code{"tls"} (total least
squares) is calculated manually via
\code{\link[base:eigen]{eigen}} and
\code{\link[stats:cov]{cov}}.}

\item{log}{a character string which contains ‘"x"’ if the x axis is
to be logarithmic, ‘"y"’ if the y axis is to be logarithmic
(base 10) and ‘"xy"’ or ‘"yx"’ if both axes are to be
logarithmic.}

\item{na.rm}{remove NA values from x and y.}

\item{verb}{verbosity level: report removed NAs?}

\item{circular}{treat data as circular, NOT fully
implemented/tested.}

\item{circular.jitter}{hack for equispaced phases (for
circular=TRUE), adds jitter to x or y (arguments
circular.jitter=c('x','y')) to avoid NA in cor.circular.}

\item{cor.legend}{plot correlation, p-value and slope (TLS) as a
legend.}

\item{line.legend}{plot line fit method as a legend.}

\item{title}{plot correlation parameters (as in legend) on the top
of the plot.}

\item{line.col}{colors of the regression lines drawn with
\code{line.methods}.  ## @param line.type line types of the
regression lines drawn with ## \code{line.methods}.}

\item{pch}{point symbols.}

\item{cex}{point size.}

\item{legpos}{position of the legend.}

\item{legbg}{background color of the legend, default: transparent
via alpha=0.  ## @param col color(s) of plotted points, if
\code{density=FALSE}, or ## color palette function if
\code{density=TRUE} (argument ## \code{colf} to
\code{\link{dense2d}}).}

\item{legcex}{font size (cex) of the legend.}

\item{signif}{number of digits to be shown for p-values in the plot
legend (argument \code{digits} to \code{\link{signif}}.}

\item{round}{number of decimal places to be shown for correlation
values in the plot legend (argument \code{digits} to
\code{\link{round}}.}

\item{density}{indicate local point densities by using
\link{dense2d} instead of the base \code{\link{plot}} function.}

\item{col}{point color if \code{density==FALSE}.}

\item{...}{further arguments to the plotting function;
\code{\link{plot}} or, if \code{density=TRUE},
\code{\link{dense2d}}.}
}
\description{
Scatter plot with correlation statistics.
}
