% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotCor}
\alias{plotCor}
\title{scatter plot with correlation statistics}
\usage{
plotCor(
  x,
  y,
  outliers,
  cor.method = c("pearson", "kendall", "spearman"),
  line.methods = c("ols", "tls"),
  na.rm = TRUE,
  circular = FALSE,
  cor.legend = TRUE,
  line.col = c(1, 2),
  pch = 20,
  cex = 1,
  legpos,
  signif = 1,
  round = 1,
  density = TRUE,
  ...
)
}
\arguments{
\item{x}{the x coordinates of the points in the plot.}

\item{y}{the y coordinates of the points in the plot.}

\item{outliers}{vector of indices or logical vector of x,y values to
exclude from correlation analysis.}

\item{cor.method}{method to calculate correlation and p-value via
\code{\link[stats:cor.test]{cor.test}}.}

\item{line.methods}{regression line methods, where \code{"ols"} is
a linear regression (ordinary least squares) with R's
\code{\link[stats:lm]{lm}} and \code{"tls"} (total least squares)
is calculated manually via \code{\link[base:eigen]{eigen}} and
\code{\link[stats:cov]{cov}}.}

\item{na.rm}{remove NA values from x and y.}

\item{circular}{NOT fully implement, treat data as circular.}

\item{cor.legend}{plot correlation, p-value and slope (TLS) as a legend.}

\item{line.col}{colors of the regression lines drawn with
\code{line.methods}.}

\item{pch}{point symbols.}

\item{cex}{point size.}

\item{legpos}{position of the legend.}

\item{signif}{number of digits to be shown for p-values in the plot
legend (argument \code{digits} to \code{\link{signif}}.}

\item{round}{number of decimal places to be shown for correlation
values in the plot legend (argument \code{digits} to
\code{\link{round}}.}

\item{density}{indicate local point densities by using
\link{dense2d} instead of the base \code{\link{plot}} function.}

\item{...}{further arguments to the plotting function;
\code{\link{plot}} or, if \code{density=TRUE},
\code{\link{dense2d}}.}
}
\description{
scatter plot with correlation statistics
}
