% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{plotdev}
\alias{plotdev}
\title{Switch between plot devices}
\usage{
plotdev(file.name = "test", type = "png", width = 5, height = 5,
  res = 100, bg = "white")
}
\arguments{
\item{file.name}{file name without suffix (.png, etc)}

\item{type}{plot type: png, jpeg, eps, pdf, tiff or svg}

\item{width}{figure width in inches}

\item{height}{figure height in inches}

\item{res}{resolution in ppi (pixels per inch), only for 'png' and 'tiff'}

\item{bg}{background color}
}
\description{
Switch between plot devices
}
