% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotDFT}
\alias{plotDFT}
\title{plot polar coordinates}
\usage{
plotDFT(dft, cycle = 3, col, ...)
}
\arguments{
\item{dft}{the Fourier transform of a time series as returned
by \code{t(mvfft(t(timeseries)))}, or alternatively, a `timeseries' object
from segmenTier's
\code{\link[segmenTier:processTimeseries]{processTimeseries}} when
run with (\code{use.fft=TRUE})}

\item{cycle}{the number of cycles (index of non-DC DFT component)
to be plotted}

\item{col}{a color vector for the rows in argument \code{dft} or
alternatively,  `clustering' object as returned by
segmenTier's \code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}
with coloring information}

\item{...}{arguments to the base \code{\link[graphics:plot]{plot}} function}
}
\description{
Plots the components of a Discrete Fourier Transform (DFT)
as polar coordinates (Re and Im of the complex numbers in the DFT).
Arguments \code{dft} and \code{col} can be segmenTier timeseries
and clustering objects.
}
