% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotOverlaps}
\alias{plotOverlaps}
\title{plot cluster-cluster overlaps}
\usage{
plotOverlaps(x, p.min = 0.01, n = 100, ...)
}
\arguments{
\item{x}{a `clusterOverlaps' object returned by
\code{\link{clusterCluster}}}

\item{p.min}{significance cutoff, p-values equal or smaller to
this cutoff will appear black; TODO: plot legend}

\item{n}{number of gray shades between \code{p=1} (white)
and \code{p >= p.min} (black)}

\item{...}{arguments to \code{\link{image_matrix}}}
}
\description{
plots the significance distribution of a cluster-cluster
overlap statistics provided by \code{\link{clusterCluster}}
}
