% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotOverlaps}
\alias{plotOverlaps}
\title{plot cluster-cluster or segment-segment overlaps}
\usage{
plotOverlaps(
  x,
  p.min = 0.01,
  p.txt = p.min * 5,
  n = 100,
  col,
  values = c("overlap", "statistic", "jaccard", "intersect.target", "text"),
  type = 1,
  txt.col = c("black", "white"),
  rmz = TRUE,
  short = TRUE,
  scale = 1,
  round,
  axis = 1:2,
  show.sig = TRUE,
  show.total = FALSE,
  ...
)
}
\arguments{
\item{x}{a `clusterOverlaps' object returned by
\code{\link{clusterCluster}}}

\item{p.min}{significance cutoff, p-values equal or smaller to
this cutoff will appear black (one-sided tests) or red/blue
(two-sided tests)}

\item{p.txt}{p-value cutoff for showing overlap numbers as white instead
of black text}

\item{n}{number of color shades between \code{p=1} (white)
and \code{p >= p.min} (black)}

\item{col}{color ramp, default are grey values (one-sided) or
red (above) and blue (below) for two-sided tests, length of
this vector overrules parameter \code{n}}

\item{values}{selection of text (numeric values) to plot, depends
on available data in \code{x}, see "Description"}

\item{type}{1 for one-sided or 2 for two-sided tests color scheme,
negative values of two-sided tests are reflected in negative p-values}

\item{txt.col}{two colors used for the plot text, ie., the
overlap counts; the second is used if `p<p.txt` as a discrete
signficance cutoff}

\item{rmz}{remove 0 from text values}

\item{short}{logical, indicating whether to cut higher overlap
numbers; currently: division by 1000 and replacement by \code{k}}

\item{scale}{factor to divide overlap numbers with, useful for
low numbers in Jaccard index}

\item{round}{number of digits to round overlap numbers to (useful
for Jaccard index)}

\item{axis}{integer vector, sets whether x-axis (1,3) and/or
y-axis (2,4) are drawn; the column and row names of \code{dat} will
be used as tick labels}

\item{show.sig}{only for overlap lists sorted by
\code{\link[segmenTier:sortClusters]{sortClusters}} from package
\code{segmenTier}:
draws a red line where unsorted non-significant hits start}

\item{show.total}{show total numbers (counts) of overlapping features
on top and right axes}

\item{...}{arguments to \code{\link{image_matrix}}}
}
\description{
Plots the significance distribution of cluster-cluster or segment-segment
overlap statistics provided by \code{\link{clusterCluster}},
\code{\link{clusterProfile}} or \code{\link{segmentJaccard}}, where
a color gradient is
calculated from \code{-log(p)}, and the text shows the overlap numbers,
e.g., the number of overlapping features for \code{\link{clusterCluster}},
or the Jaccard index or relative intersect values for
\code{\link{segmentJaccard}}. Option \code{text} allows to select
which values to plot as text. Only "overlap" is available for a
\code{\link{clusterCluster}} results, while for \code{\link{segmentJaccard}}
the Jaccard index ("jaccard"), or the relative intersect size can
be shown: "intersect.target" and "intersect.query" are the intersect
divided by total target or query length, respectively.
Note that two
distinct p-value cutoffs can be visualyized: p-values \code{<=p.min}
are shown in black, and p-values \code{p.txt} are shown in white instead
of black text (thus becoming visible on the black of significant
overlaps).
}
