% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{plotSingles}
\alias{plotSingles}
\title{plot indivividual time-courses}
\usage{
plotSingles(x, cls, goi, each, lwd = 2, leg.xy = "topleft", ...)
}
\arguments{
\item{x}{"timeseries" object from function
\code{\link[segmenTier:processTimeseries]{processTimeseries}}}

\item{cls}{"clustering" object (simple list)}

\item{goi}{list of feature ids (rownames in cls$clusters) to plot}
}
\description{
plot indivividual time-courses (GOI: genes of interest) from
timeseries and clustering object (simple list)
}
