% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R, R/segmenTools.R_selectColors.R
\name{pvalDist}
\alias{pvalDist}
\title{cluster p-value summary}
\usage{
pvalDist(pvs, threshold = 0.01)

pvalDist(pvs, threshold = 0.01)
}
\arguments{
\item{pvs}{vector of p-values}

\item{threshold}{p-value threshold to be counted as signficant}

\item{pvs}{vector of p-values}

\item{threshold}{p-value threshold to be counted as signficant}
}
\description{
calculates the mean of p-values and number of significant
p-values (<\code{threshold}) and number of NA values

calculates the mean of p-values and number of significant
p-values (<\code{threshold}) and number of NA values
}
