% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readFASTA}
\alias{readFASTA}
\title{fasta sequence file parser stolen from (an older version of)
\code{Biostrings}}
\usage{
readFASTA(file, checkComments = TRUE, strip.descs = TRUE,
  split.desc = "")
}
\arguments{
\item{file}{file with sequence(s) in fasta format}

\item{checkComments}{remove information after a semi-colon ; in fasta
header}

\item{strip.descs}{split header at a pattern in argument \code{split.desc}
and store in returned structure as \code{details}}

\item{split.desc}{pattern to split header into description and details
with argument \code{strip.descs}}
}
\description{
fasta sequence file parser stolen from (an older version of)
\code{Biostrings}
}
