% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coor2index.R
\name{removeCircularFeatures}
\alias{removeCircularFeatures}
\title{NOT WORKING - Undo \code{expandCircularFeatures}
searches for circular features by the \code{idTag} added
to ID and type columns in \code{expandCircularFeatures},
and maps downstream coordinates back to original features.}
\usage{
removeCircularFeatures(features, idTag = "-circ2", idCols = c(ID = "ID",
  type = "type", parent = "parent"), rmParent = TRUE)
}
\arguments{
\item{features}{list of genomic features with coordinates}

\item{idTag}{tag used for tagging downstream halves}

\item{idCols}{named vector of column names for feature ID, type,
and feature parent; note that a "parent" column will be removed if
it is (a) empty and (b) argument \code{rmParent==TRUE}}

\item{rmParent}{the column parent}
}
\description{
NOT WORKING - Undo \code{expandCircularFeatures}
searches for circular features by the \code{idTag} added
to ID and type columns in \code{expandCircularFeatures},
and maps downstream coordinates back to original features.
}
\seealso{
\code{expandCircularFeatures}
}
