% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\name{segmentJaccard_bed}
\alias{segmentJaccard_bed}
\title{calculate segment overlap statistics using UCSC bedtools}
\usage{
segmentJaccard_bed(
  query,
  target,
  qclass,
  tclass,
  prefix = "cl_",
  chrL,
  perm,
  tmpdir,
  runid,
  save.permutations = FALSE,
  verb = 1
)
}
\arguments{
\item{query}{query segment table, will be permutated.}

\item{target}{target segment table.}

\item{qclass}{column name which holds a sub-classification (clustering) of
the query segments, omit to use all.}

\item{tclass}{column name which holds a sub-classification (clustering) of
the target segments, omit to use all.}

\item{prefix}{prefix to be added to the name and score columns of bed
files (column 4,5), used to get unique names.}

\item{chrL}{obligatory vector of chromosome length, in the order
used as index in the \code{chr} column of query and target.}

\item{perm}{number of permutations for calculating statistics.}

\item{tmpdir}{temporary directory, useful to provide for debugging,
or exploring detailed results.}

\item{runid}{use this ID for the run, for more recognizable jobnames.
This is potentially dangerous since each call MUST have a unique name.}

\item{save.permutations}{save permutated query bed files used
in p-value calculation. If this is provided together with
 \code{tmpdir}, the randomized bed files can be re-used. However, this
may be dangerous or cause errors when different queries and targets
are used with the same \code{tmpdir}.}

\item{verb}{verbosity level, 0: silent.}
}
\description{
This function produces the same output as \code{\link{segmentJaccard}},
but using a command-line call to UCSC bedtools, which must be
installed, and starting from normal genome coordinates (w/o indexing;
0-based and inclusive start and end). It temporarily creates bed files,
and can require quite a lot of disk space. However, it is faster
than \code{\link{segmentJaccard}}, and likely better tested.
}
