% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenTools.R
\docType{package}
\name{segmenTools}
\alias{segmenTools}
\alias{segmenTools-package}
\title{segmenTools : analysis of genome segmentations by segmenTier}
\description{
segmenTools : analysis of genome segmentations by segmenTier
}
\section{Dependencies}{
 basic (\code{stats}, \code{graphics}, \code{grDevices}), clustering, \code{flowClust}, \code{flowMerge}
}

\author{
Rainer Machne \email{raim@tbi.univie.ac.at}
}
