% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{selected}
\alias{selected}
\title{get clustering ID}
\usage{
selected(cset, K, name = TRUE)
}
\arguments{
\item{cset}{a structure of class 'clustering' as returned by
segmenTier's \code{\link[segmenTier:clusterTimeseries]{clusterTimeseries}}}

\item{K}{cluster number; if missing the `selected' clustering will
be taken}

\item{name}{logical, if TRUE the name of of the selected clustering
will be returned, if FALSE its index number}
}
\description{
get the column name or index of the selected
clustering; allows to interface values for the selected clustering
}
