% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{sortOverlaps}
\alias{sortOverlaps}
\title{sorts cluster overlap structure by p-values}
\usage{
sortOverlaps(ovl, p.min = 0.05, axis = 2, cut = FALSE)
}
\arguments{
\item{ovl}{a `clusterOverlaps' object returned by
\code{\link{clusterCluster}}}

\item{p.min}{significance cutoff during sorting}

\item{axis}{axis to sort (2 for y-axis/rows, 1 for x-axis/columns)}

\item{cut}{remove all overlaps without any \code{p<p.min}}
}
\description{
Sorts one dimension of a cluster overlap structure
by their p-values along the sorting of the other axis.
For each cluster along the pre-sorted (non-selected) axis,
the most significant overlaps (\code{p<p.min}) are chosen
and moved to the top of the matrix.
TODO: align axis selection with nomenclature in clusterCluster
}
