% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTools.R
\name{sortOverlaps}
\alias{sortOverlaps}
\title{sorts cluster overlap structure by p-values}
\usage{
sortOverlaps(ovl, axis = 2, p.min = 0.05, cut = FALSE, srt)
}
\arguments{
\item{ovl}{a `clusterOverlaps' object returned by
\code{\link{clusterCluster}}}

\item{axis}{axis to sort (2 for y-axis/rows, 1 for x-axis/columns)}

\item{p.min}{significance cutoff during sorting}

\item{cut}{remove all overlaps without any \code{p<p.min}}

\item{srt}{sorting vector for rows, if this is passed the significance
sorting is skipped}
}
\description{
Sorts one dimension of a cluster overlap structure
by their p-values along the sorting of the other axis.
For each cluster along the pre-sorted (non-selected) axis,
the most significant overlaps (\code{p<p.min}) are chosen
and moved to the top of the matrix.
}
