% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{summarizeGEOSoft}
\alias{summarizeGEOSoft}
\title{summarize GEOSoft probes}
\usage{
summarizeGEOSoft(data, id = "ORF", avg = "mean", farms = FALSE,
  keep.empty = FALSE, replicate = FALSE, repsep = ";", verb = TRUE, ...)
}
\arguments{
\item{data}{a list as returned by function \code{\link{parseGEOSoft}}}

\item{id}{a column ID; probes with equal strings in this column will
be summarized}

\item{avg}{a function (or the name of a function as a string)
for calculating an `average' value for each cluster; default is
the \code{mean}}

\item{farms}{logical to indicate to use the bioconductor package
\pkg{farms} for probe summarization (function \code{\link[farms:generateExprVal.method.farms]{generateExprVal.method.farms}})}

\item{keep.empty}{keep columns with no entry in column ID by
copying the probe ID}

\item{replicate}{if TRUE probe sets assigned to multiple features
(separated by string provided in rep.sep) will be duplicated to
and each feature will get its own row}

\item{repsep}{separator for multiple-feature probes}

\item{verb}{print messages}

\item{...}{arguments to the function specified in \code{avg} or for
function \code{\link[farms:generateExprVal.method.farms]{generateExprVal.method.farms}} if \code{farms==TRUE}}
}
\description{
takes as input the list returned by \code{\link{parseGEOSoft}},
a GEO microarray data set, including a table of probe mappings,
as provided via GEO soft family files.
Argument \code{id} specifies a probe mapping column ID.
Multiples probes for the same features (same string in column ID) will
then be summarized by the function in argument \code{avg}
(default: median). If the package \pkg{farms} installed, the more
sophisticated probe summarization of this package can be used
\url{https://www.bioconductor.org/packages/release/bioc/html/farms.html}.
}
