% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{val2col}
\alias{val2col}
\title{Map values to colors}
\usage{
val2col(x, cols, limits = NULL)
}
\arguments{
\item{x}{numeric values}

\item{cols}{a vector of colors}

\item{limits}{min/max cut-off values of x}
}
\description{
A simple function copied from
}
